### 2022-02-14 #4160 @whedon
[PRE REVIEW]: FEM_2D: A Rust Package for 2D Finite Element Method Computations with Extensive Support for *hp*-refinement
**Submitting author:** @jeremiah-corrado (<a href="http://orcid.org/0000-0003-2688-0600">Jeremiah Corrado</a>)
**Repository:** <a href="https://github.com/jeremiah-corrado/fem_2d" target ="_blank">https://github.com/jeremiah-corrado/fem_2d</a>
**Version:** 0.1.0
