### 2022-02-14 79295bc
fixed DOIs in paper.bib
