+++ paper.bib
@@ -28,12 +28,15 @@
   url = {https://github.com/dimforge/nalgebra}
 }
 
-@misc{slepc,
-  title = {SLEPc: Scalable Library for Eigenvalue Problem Computations},
-  year = {2021},
-  publisher = {GitHub},
-  journal = {GitHub repository},
-  url = {https://github.com/firedrakeproject/slepc}
+@article{slepc,
+   author  = "Vicente Hernandez and Jose E. Roman and Vicente Vidal",
+   title   = "{SLEPc}: A scalable and flexible toolkit for the solution of eigenvalue problems",
+   journal = "ACM Trans. Math. Software",
+   volume  = "31",
+   number  = "3",
+   pages   = "351--362",
+   year    = "2005"
+   DOI  = {10.1145/1089014.1089019}
 }
 
 @misc{petsc-web-page,
@@ -72,6 +75,7 @@
   publisher = {Birkh{\"{a}}user Press},
   pages     = {163--202},
   year      = {1997}
+  DOI = {10.1007/978-1-4612-1986-6_8}
 }
 
 @article{dealII93,
