### 2022-02-14 0e8e954
fixed DOIs in paper.bib
