+++ paper.bib
@@ -35,7 +35,7 @@
    volume  = "31",
    number  = "3",
    pages   = "351--362",
-   year    = "2005"
+   year    = "2005",
    DOI  = {10.1145/1089014.1089019}
 }
 
@@ -74,7 +74,7 @@
   editor    = {E. Arge and A.~M. Bruaset and H.~P. Langtangen},
   publisher = {Birkh{\"{a}}user Press},
   pages     = {163--202},
-  year      = {1997}
+  year      = {1997},
   DOI = {10.1007/978-1-4612-1986-6_8}
 }
 
