### 2022-02-16 #4160 @jedbrown
This looks cool and I'm happy to take it. @jeremiah-corrado I didn't see [community guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html#community-guidelines) in your documentation. Could you add those to your README and/or [standard templates](https://github.com/jeremiah-corrado/fem_2d/community)? I'll line up reviewers.
