+++ README.md
@@ -21,10 +21,8 @@ A Rust library for 2D Finite Element Method computations, featuring:
 
 ## Usage
 
-Add this to your `Cargo.toml` file:
-
+Include this line in your `Cargo.toml` file under **\[dependencies\]**:
 ```toml
-[dependencies]
 fem_2d = "0.1.0"
 ```
 
@@ -32,6 +30,10 @@ Please include one or more of the following citations in any academic or commerc
 * ***metapaper*** (comming soon!)
 * Corrado, Jeremiah; Harmon, Jake; Notaros, Branislav (2021): A Refinement-by-Superposition Approach to Fully Anisotropic hp-Refinement for Improved Efficiency in CEM. TechRxiv. Preprint. [https://doi.org/10.36227/techrxiv.16695163.v1](https://doi.org/10.36227/techrxiv.16695163.v1)
 * Harmon, Jake; Corrado, Jeremiah; Notaros, Branislav (2021): A Refinement-by-Superposition hp-Method for H(curl)- and H(div)-Conforming Discretizations. TechRxiv. Preprint. [https://doi.org/10.36227/techrxiv.14807895.v1](https://doi.org/10.36227/techrxiv.14807895.v1)
+
+## Documentation
+
+The latest Documentation can be found [here](https://docs.rs/fem_2d/0.1.0/fem_2d/)
 
 ## Example
 
@@ -124,6 +126,14 @@ let u_plus_1_v_minus_3 = PRef::from(1, -3);
 
 
 ## JSON Mesh Files
+
+fem_2d uses `.json` files to import and export Mesh layouts. 
+
+* The input format is simplified and describes the geometry of the problem only. 
+* The output format describes the Mesh in it's refined state which is usefull for debugging and observing the refinement state.
+
+### Input Mesh Files
+
 A `Mesh` can be constructed from a JSON file with the following format:
 ```JSON
 {
@@ -171,16 +181,23 @@ Both Elements start with a polynomial expansion order of 1 upon construction.
 
 This library does not yet support curvilinear elements. When that feature is added, this file format will also be extended to describe higher-order geometry.
 
-___
+### Output Mesh Files
 
 A refined `Mesh` can also be exported and visualized using [this](https://github.com/jeremiah-corrado/fem_2d_mesh_plot) tool:
 
 <img src="/rm_figs/mesh_example.jpeg" width="350">
 
-
----
+## Solution Plotting
 
 Solution figures like the ones below can be generated by exporting solutions as `.vtk` files and plotting using a compatible tool (such as [Visit](https://visit-dav.github.io/visit-website/index.html)):
 
 <img src="/rm_figs/e_mag_1.jpeg" width="400">
 <img src="/rm_figs/e_mag_9.jpeg" width="400">
+
+## Code of Conduct
+
+
+
+Contributions in any form (issues, pull requests, etc.) to this project must adhere to Rust's [Code of Conduct](https://www.rust-lang.org/policies/code-of-conduct).
+
+Additionally, all contributions to this library will be subject to the MIT liscence. 
