+++ rm_figs/contributing.md
@@ -1,0 +1,1 @@
+# Contributing to fem_2d
