+++ README.md
@@ -194,10 +194,11 @@ Solution figures like the ones below can be generated by exporting solutions as 
 <img src="/rm_figs/e_mag_1.jpeg" width="400">
 <img src="/rm_figs/e_mag_9.jpeg" width="400">
 
-## Code of Conduct
+## Community Guidelines
 
+Contributions, questions, and bug-reports are welcome! Any pull requests, issues, etc. must adhere to Rust's [Code of Conduct](https://www.rust-lang.org/policies/code-of-conduct).
 
+More details about how to contribute to this project can be found [here](https://github.com/jeremiah-corrado/fem_2d/blob/main/rm_figs/contributing.md).
 
-Contributions in any form (issues, pull requests, etc.) to this project must adhere to Rust's [Code of Conduct](https://www.rust-lang.org/policies/code-of-conduct).
+Questions and bug-reports should be directed to the Issues tab. 
 
-Additionally, all contributions to this library will be subject to the MIT liscence. 
