+++ README.md
@@ -194,7 +194,7 @@ Solution figures like the ones below can be generated by exporting solutions as 
 <img src="/rm_figs/e_mag_1.jpeg" width="400">
 <img src="/rm_figs/e_mag_9.jpeg" width="400">
 
-## Community Guidelines
+## Community Guidelines / Code of Conduct
 
 Contributions, questions, and bug-reports are welcome! Any pull requests, issues, etc. must adhere to Rust's [Code of Conduct](https://www.rust-lang.org/policies/code-of-conduct).
 
