### 2022-02-16 #4172 @whedon
[REVIEW]: FEM_2D: A Rust Package for 2D Finite Element Method Computations with Extensive Support for *hp*-refinement
**Submitting author:** @jeremiah-corrado (<a href="http://orcid.org/0000-0003-2688-0600">Jeremiah Corrado</a>)
**Repository:** https://github.com/jeremiah-corrado/fem_2d
**Branch with paper.md** (empty if default branch):
**Version:** 0.2.2
