### 2022-03-11 #4172 @jedbrown
Whedon retired young so we now ask `@editorialbot generate my checklist`. You don't need an invite.
