### 2022-03-19 jeremiah-corrado/fem_2d#2 @YohannDudouit
Observations and suggestions for the JOSS review.
Here are some observations and suggestions that you're welcome to consider:

Overall I think the code could push further the separation between abstraction and implementation, and provide a few more building blocks for more general finite element problems. For instance, the `BasisFn` struct seem much closer to implementation details than abstracting the mathematical concept of a `basis`. Many parts of the library are undocumented, making it difficult to grasp the details of the library. The module structure could also better represent the structure of the project, it can be difficult to navigate and grasp the organization of the library, for instance `fem_2d::basis::glq` and `fem_2d::integration::integrals::glq` why are these separated?

## On the structure/organization/documentation/naming:

- The first module layer is: `basis`, `domain`, `integration`, `linalg`, `prelude`. I think the structure would gain in clarity with a higher level of module of the type `mesh`, `fem`, `linalg`.

- The `basis` module contains the `glq` (Gauss-Legendre Quadrature points and weights), it seems that numeric integration rules would better fit in the `integration` module.

- The methods of the `ShapeFn` trait are undocumented, as a result it is hard to implement this trait for any custom type.

- The `BasisFn` struct has too little documentation for a user to understand how to use this structure.

- What is a `BasisFnSampler`?

- The names "basis" and "shape" are used in a way that can be confusing to users, `ShapeFn` does not seem to behave as a function, and `BasisFn` does not seem to be a collection of shape functions.

- The names `Elem` and `Element` would probably gain in clarity if renamed something like `ReferenceElement` and `PhysicalElement`.

- `mesh` contains the module `edge`, `elem`, `element`, `h_refinement`, `node`, `p_refinement`, and `space`, it seems that this module conflates two notions usually separated: `mesh` and `finite element space`. `edge`, `node`, and `space::Point` should probably be in `mesh`; whereas `elem`, `element`, `h_refinement`, `p_refinement` could probably be in a `finite_element_space` module.

- `fem_2d::domain::mesh::space` contains the structs `M2D` and `V2D` (these types would also benefit from more self-documented names), these seem out of place and would probably better fit somewhere in the `linalg` module.

## On the design:

- The `Mesh` struct feels more like a `FiniteElementSpace` or `FiniteElementDomain` struct, I would suggest renaming and introducing another `Mesh` struct focusing only on cells, faces, edges, and nodes.

- The `dof` module is confusing, degrees of freedom are usually scalar values associated to basis functions in order to describe a finite element field. It seems that a `Dof` contains a lot of information that I would expect to find in the `BasisFn`.

- `linalg` is it the role of a finite element library to wrap linear algebra, I would find the opposite to be more idiomatic, i.e., implementing linear algebra traits for finite element operators to be usable by linear algebra libraries.

- The `ShapeFn` trait seems over specialized: "Hierarchical Curl-Conforming Shape Function along a single direction (defined over (-1.0, +1.0)). A Custom Basis Function can be defined by implementing this Trait on your own Struct." I think renaming this trait `HierarchicalCurlShapeFn` could be more aligned with the documentation, and adding base traits for shape functions without Curl-conformity nor hierarchical constraints could make sense:
```rust
pub trait ShapeFn: Clone + Sync + Send + Debug {
   ...
}

pub trait CurlShapeFn: ShapeFn { ... }

pub trait HierarchicalShapeFn: ShapeFn { ... }

pub trait HierarchicalCurlShapeFn: CurlShapeFn + HierarchicalShapeFn { ... }
```
- It is not clear how the Curl-conformity contract is verified, is it through the interface? Same for the hierarchical property.

- I think the pattern `Object`/`ParObject` is not very idiomatic, two implementations of the same concept (same semantic) should not result in two disjoint types. I think they should either be unified behind a trait, or their parallel policy should become a type parameter `Object<SerialPolicy/ParallelPolicy>`.

- The `Materials` struct seems to be very specialized to problems where each element's physical properties can be described with two scalar constants per element, this seems very limited to simple Maxwell problems. Is it easy for a user to use its own physical description? How would that work if physical properties were to be defined with a function inside each element?

## On the implementation:

- The only type implementing `ShapeFn` is `KOLShapeFn`, what does "KOL" stands for? The documentation could be more extensive: "A simple Curl-Conforming Hierarchical Shape Function".

- It would be helpful for the user to implement one or some of the classic non-Curl basis functions: GL, GLL, Bernstein as a reference implementation.

- The only integral operator defined is `curl_curl`, it would help new users implement their own integrals if classic operators like a mass operator `< u, v >`, and Poisson operator `< ∇u, ∇v >` were also implemented.

- Would it be possible to have access to some linear solvers, these also seem essential for finite element analysis.
