### 2022-03-22 #4172 @YohannDudouit
I opened the issue https://github.com/jeremiah-corrado/fem_2d/issues/2 containing a number of observations and suggestions.
