### 2022-03-24 jeremiah-corrado/fem_2d#3 @jeremylt
Unified Error Handling Strategy
Yohann has covered many good design points, so I will focus on some more Rust specific implementation details.

In my experience, Rust users tend to expect error handling to work like:
```rust
pub fn this_fn_could_fail(some_arg: usize) -> MyCrateResult<usize> {
  // ? to propagate possible error from function call
  let some_intermediate_value = other_possible_to_fail(some_arg)?;
  Ok(42 * some_intermediate_value)
}
```

You seem to have a mix of error handling strategies in this crate that isn't immediately clear to me. The `unwrap()` method is called on intermediate results from function calls internally that seem like they could fail.

I recommend a) unifying your error handling strategy or b) documenting why you have different error handling strategies in different parts of the codebase.

Similarly, in the paper, I would update the example to use `?` instead of `.unwrap()` - from what I've seen, its currently a bit more idiomatic for the main driver function to return `Result<()>`.
