### 2022-03-24 jeremiah-corrado/fem_2d#4 @jeremylt
Struct Public Members
It seems like many structs have `pub` fields that are not used in the examples. Do you intent for users to have access to all of this data, or could some of it be made `pub(crate)`?

For example, the `BasisFn` struct exposes

>t: Raw transformation matrices at each sample point. Describes transformation from real space to sampled parametric space.

The word 'raw' makes me think this is perhaps something that users should not have access to. I try to make as many fields as possible `pub(crate)` over `pub` so I have freedom in future refactors. Anything declared `pub` that is changed may result in more invasive changes than intended in user code.

I recommend double checking which struct fields need to be `pub` and which can be made `pub(crate)`.
