### 2022-03-24 jeremiah-corrado/fem_2d#5 @jeremylt
Documentation Code Samples
Rust supports code examples in the doc comments. I find that this tends to be expected by Rust users. Also, these code examples can be run as part of your test suite.

https://doc.rust-lang.org/rust-by-example/meta/doc.html

I recommend including inline code examples in the doc comments, especially for your most key functions that you expect users to interact with the most. I would also run these code examples as part of your test suite.

You can run the code in the README as part of your test suite by including it in the crate level doc comments by using this line in `src/lib.rs`
```rust
#![doc = include_str!("../README.md")]
```
This can reduce duplication and ensure that your example code is up to date.
