### 2022-03-24 jeremiah-corrado/fem_2d#6 @jeremylt
Continuous Integration Testing
I see a job here for building the paper, but not for running the test suite:

https://github.com/jeremiah-corrado/fem_2d/tree/main/.github/workflows

[`actions-rs`](https://github.com/actions-rs) provides a wide range of good GitHub actions for [testing](https://github.com/actions-rs/cargo) and [code coverage](https://github.com/actions-rs/tarpaulin) that make CI fairly painless.

I recommend including a CI job for running your crate test suite. I would also consider using Tarpaulin, possibly in CI, to explore areas where bugs may be hiding in your code due to low or no testing of various functions.
