### 2022-03-24 #4172 @jeremylt
Since @YohannDudouit made a pretty comprehensive review on the overall structure and I agree with many of his points, I focused on creating a few issues highlighting some more Rust specific technical aspects:

* [Error Handling](https://github.com/jeremiah-corrado/fem_2d/issues/3)
* [Struct Member Access](https://github.com/jeremiah-corrado/fem_2d/issues/4)
* [Doc Comment Code Examples](https://github.com/jeremiah-corrado/fem_2d/issues/5)
* [CI Tests](https://github.com/jeremiah-corrado/fem_2d/issues/6)
