### 2022-03-24 #4172 @jeremiah-corrado
Thank you both for the detailed feedback! It's really nice to get an outside perspective on the design and documentation.

I am currently in prep-mode for my thesis defense, but I should be able to start addressing these issues within a couple of weeks.

In the meantime, I have a few questions and notes about the suggestions you made:

@YohannDudouit:
- I agree that much of the conceptual partitioning could use some slight adjustment, as well as better naming and documentation; however, I think a few of the structures should stay where they are. For example, M2D and V2D should be in the space module as part of the broader mesh module because they describe points and transformations in the 2D physical/parametric spaces where elements are defined. These are specialized data structures that do not have any utility to the linear algebra portion of the crate. Would it make more sense to rename the linalg module to something like "linear-solvers" to avoid this confusion?
- The ShapeFn trait is somewhat specialized to the work I've been doing with the Maxwell Eigenvalue problem. I like your idea about building up a hierarchy of traits to support a wider and more abstract range of shape function types! I'll look into this a bit further.
- MaxOrthoBasisFn also implements the ShapeFn trait, but it has to be included with the Nightly compiler using a feature-flag. I'll include more documentation about this.
- There is a [Mass< u, v > operator](https://docs.rs/fem_2d/latest/fem_2d/integration/integrals/inner/index.html)
- Generally, I am not sure where to draw the line between implementing more functionality myself, and inviting open-source contribution for some of these features, such as extended ShapeFns, Integrals, and a linear-solver. Do you think it would be appropriate to leave some of these things as open issues for people to work on? Or should they be part of the library right away?

@jeremylt:
- I was not aware of Tarpaulin, so I'll give that a try.
- I also did not know about the ```#![doc = include_str!("../README.md")]``` flag. I'll include that to take care of crate-level testing and then move towards more code-sample based testing (in addition to the module-level tests).
- As far as I know, the public struct members are all intended to be public. I'll provide some more concise documentation or examples of how those could be used.
- All of the internal calls to .unwrap() are made when I **know** that a function call will not return an Err. For example, in the [```global_h_refinement()```](https://docs.rs/fem_2d/latest/fem_2d/domain/mesh/struct.Mesh.html#method.global_h_refinement) method on Mesh, I call unwrap() on [```elem_is_h_refineable()```](https://docs.rs/fem_2d/latest/fem_2d/domain/mesh/struct.Mesh.html#method.elem_is_h_refineable) because it only returns an Err when the provided elem_id does not exist. In this case I know that it does exist because I am iterating over the list of elems. Therefore, I can call unwrap rather than propagating the error using the "?" operator. Would it be helpful if I make a case for the validity of calling unwrap in the documentation for these methods?
- I'll update the paper example (and readme example) to return a Result<()> from main and use  ?  instead of  unwrap(). I agree that this is more idiomatic, I was just trying to avoid the often confusing ```Ok(())``` at the bottom of the function.
- Otherwise, did you have additional concerns about error handling? If so, would you be able to provide a specific example of the non-uniformity you are referring to?

Thanks again for all the excellent feedback!
