+++ src/domain/mesh/h_refinement.rs
@@ -4,17 +4,20 @@ use std::fmt;
 use std::ops::AddAssign;
 
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
-/// Description of an Elem's h-Refinement levels in the u and v directions
+/// Description of an `Elem`s h-Refinement levels in the u and v directions
+///
+/// These values facilitate the matching of basis functions among neighboring `Elem`s
 pub struct HLevels {
     pub u: u8,
     pub v: u8,
 }
 
 impl HLevels {
-    pub fn from(u: u8, v: u8) -> Self {
+    fn from(u: u8, v: u8) -> Self {
         Self { u, v }
     }
 
+    /// Construct a new set of [HLevels] from an [HRef] of `self`
     pub fn refined(&self, refinement: HRef) -> Self {
         match refinement {
             HRef::T => Self::from(self.u + 1, self.v + 1),
@@ -23,6 +26,7 @@ impl HLevels {
         }
     }
 
+    /// The ranking of the associated `Elem` according to the local `Edge`s (and their direction)
     pub fn edge_ranking(self, edge_dir: ParaDir) -> [u8; 2] {
         match edge_dir {
             ParaDir::U => [self.v, self.u],
@@ -30,6 +34,7 @@ impl HLevels {
         }
     }
 
+    /// The ranking of the associated `Elem` according to the local `Node`s
     pub fn node_ranking(self) -> [u8; 2] {
         [self.u, self.v]
     }
@@ -51,7 +56,15 @@ impl From<HLevels> for JsonValue {
     }
 }
 
-/// Description of an h-Refinement
+/// The h-Refinement Type
+///
+/// h-refinements are used to introduce new `Elem`s into the mesh
+/// * T: 4 new `Elem`s are introduced over a given `Elem`
+/// * U(None): 2 new `Elem`s are introduced over a given `Elem` (improving resolution in the u-direction)
+/// * V(None): 2 new `Elem`s are introduced over a given `Elem` (improving resolution in the v-direction)
+/// * U(Some( 0 or 1 )): A U-type refinement, followed by a V-type refinement of the 0th or 1st resultant `Elem`
+/// * V(Some( 0 or 1 )): A V-type refinement, followed by a U-type refinement of the 0th or 1st resultant `Elem`
+///
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
 pub enum HRef {
     /// isotropic
@@ -80,7 +93,7 @@ impl HRef {
 
     /// Extended U-type refinement
     ///
-    /// The 0th or 1st child generated by the U-Type is then V-Type refined.
+    /// The 0th or 1st child generated by the U-Type refinement is then V-Type refined
     /// Returns an error if `child_idx` is not 0 or 1
     pub const fn u_extened(child_idx: u8) -> Result<Self, HRefError> {
         match child_idx {
@@ -92,7 +105,7 @@ impl HRef {
 
     /// Extended V-type refinement
     ///
-    /// The 0th or 1st child generated by the V-Type is then U-Type refined.
+    /// The 0th or 1st child generated by the V-Type refinement is then U-Type refined
     /// Returns an error if `child_idx` is not 0 or 1
     pub const fn v_extened(child_idx: u8) -> Result<Self, HRefError> {
         match child_idx {
@@ -157,21 +170,25 @@ impl AddAssign for HRef {
                 _ => (),
             },
             HRef::U(Some(idx)) => match rhs {
-                HRef::U(Some(r_idx)) => if r_idx != *idx {
-                    *self = HRef::T;
-                },
+                HRef::U(Some(r_idx)) => {
+                    if r_idx != *idx {
+                        *self = HRef::T;
+                    }
+                }
                 HRef::V(_) => *self = HRef::T,
[…]
