+++ src/integration/glq.rs
@@ -1,0 +1,376 @@
+use nalgebra::{DMatrix, SymmetricEigen};
+
+/// 2D Gauss Legendre Quadrature integral of some function F defined over an m by n rectangular region.
+/// ```
+/// use fem_2d::integration::glq::*;
+///
+/// // define glq points over `(-1, 1)` in both directions
+/// let (u_points, u_weights) = gauss_quadrature_points(10, false);
+/// let (v_points, v_weights) = gauss_quadrature_points(10, false);
+///
+/// // compute the integral of (u^2 * v^2)
+/// let solution = real_gauss_quad(&u_weights, &v_weights, |m, n| {
+///    u_points[m].powi(2) * v_points[n].powi(2)
+/// });
+///
+/// assert!((solution - 4.0 / 9.0).abs() < 1e-12);
+///
+/// ```
+pub fn real_gauss_quad<F>(u_weights: &[f64], v_weights: &[f64], integrand: F) -> f64
+where
+    F: Fn(usize, usize) -> f64,
+{
+    let mut solution = 0.0;
+    for (m, u_w) in u_weights.iter().enumerate() {
+        let mut inner_solution = 0.0;
+        for (n, v_w) in v_weights.iter().enumerate() {
+            inner_solution += integrand(m, n) * v_w;
+        }
+        solution += inner_solution * u_w;
+    }
+    solution
+}
+
+/// 2D Gauss Legendre Quadrature integral of some function F defined over an m by n rectangular region.
+///
+/// This is the same as [real_gauss_quad] except, the outer edge (the first and last elements of 'u_weights' and 'v_weights') are ignored.
+/// Intended to be used in scenarios where BasisFns are defined for By-Parts integration but the solutions on the edges are known to be zero.
+///
+/// ```
+/// use fem_2d::integration::glq::*;
+///
+/// // define glq points over `[-1, 1]` in both directions
+/// let (u_points, u_weights) = gauss_quadrature_points(10, true);
+/// let (v_points, v_weights) = gauss_quadrature_points(10, true);
+///
+/// // compute the integral of (u^2 * v^2)
+/// let solution = real_gauss_quad_inner(&u_weights, &v_weights, |m, n| {
+///    u_points[m].powi(2) * v_points[n].powi(2)
+/// });
+///
+/// assert!((solution - 4.0 / 9.0).abs() < 1e-12);
+/// ```
+pub fn real_gauss_quad_inner<F>(u_weights: &[f64], v_weights: &[f64], integrand: F) -> f64
+where
+    F: Fn(usize, usize) -> f64,
+{
+    let mut solution = 0.0;
+    for (m, u_w) in u_weights
+        .iter()
+        .enumerate()
+        .skip(1)
+        .take(u_weights.len() - 2)
+    {
+        let mut inner_solution = 0.0;
+        for (n, v_w) in v_weights
+            .iter()
+            .enumerate()
+            .skip(1)
+            .take(v_weights.len() - 2)
+        {
+            inner_solution += integrand(m, n) * v_w;
+        }
+        solution += inner_solution * u_w;
+    }
+    solution
+}
+
+/// 1D integral over some function F, which is defined along one edge of a rectangular parametric region.
+///
+/// ```
+/// use fem_2d::integration::glq::*;
+///
+/// // define glq points over `[-1, 1]` in both directions
+/// let (u_points, u_weights) = gauss_quadrature_points(10, true);
+/// let (v_points, v_weights) = gauss_quadrature_points(10, true);
+///
+/// // compute the integral of (u^2 * v^2) along all four "edges"
[…]
