+++ src/domain/fields.rs
@@ -6,12 +6,12 @@ use std::fs::File;
 use std::io::{BufWriter, Write};
 use std::time::SystemTime;
 
-// TODO: rework UniformFieldSpace and print_to_vtk functions to support curvilinear elements
+// TODO: update UniformFieldSpace and print_to_vtk functions after curvilinear elements are implemented
 // TODO: implement a constant (over x and y) density FieldSpace structure which supports field image exports
 
-/// A collection of Field Solutions over a [Domain].
+/// A collection of Field Solutions over a [Domain]
 ///
-/// Solutions can be operated on and printed to VTK files for visualization.
+/// Solutions can be operated on and printed to VTK files for visualization
 pub struct UniformFieldSpace<'d> {
     quantities: HashMap<String, FieldQuantity>,
     parametric_points: [Vec<f64>; 2],
@@ -22,8 +22,8 @@ pub struct UniformFieldSpace<'d> {
 impl<'d> UniformFieldSpace<'d> {
     /// Generate a FieldSpace over a [Domain]
     ///
-    /// * `domain`: is a reference to a Domain which must outlive this Structure. It is used for all subsequent computations
-    /// * `densities`: the number of evaluation points in the u and v directions respectively. These values apply to Elems on the shell of the Domain (their ancestor Elems will have more evaluation points)
+    /// The `densities` argument defines the size of the points grid that should be generated on the leaf-`Elems` (the most h-refined `Elem`s without children).
+    /// Non-leaf-`Elem`s will have denser point-grids because they are overlapped by multiple leaf-`Elem`s.
     pub fn new(domain: &'d Domain, densities: [usize; 2]) -> Self {
         Self {
             quantities: HashMap::new(),
@@ -38,19 +38,35 @@ impl<'d> UniformFieldSpace<'d> {
 
     // TODO: add option to include z-directed fields (after W-Dir & node-type Basis functions are implemented)
 
-    /// Use an eigenvector and associated [ShapeFn] to compute the X and Y fields over the `Domain`
+    /// Use an eigenvector and associated [ShapeFn] to compute the X and Y fields over the [Domain]
     ///
-    /// The X and Y field quantities will be stored as X_{vector_name} and Y_{vector_name} respectively. The Names are returned in an array in that order.
+    /// The X and Y field quantities will be stored as {vector_name}_x and {vector_name}_y respectively. The Names are returned in an array in that order.
+    ///
+    /// ```
+    /// use fem_2d::prelude::*;
+    ///
+    /// let domain = Domain::unit();
+    /// let unit_solution = vec![1.0; domain.dofs.len()];
+    ///
+    /// // construct a field space with a 10x10 grid on each leaf-`Elem`
+    /// let mut ufs = UniformFieldSpace::new(&domain, [10, 10]);
+    ///
+    /// // compute the X and Y fields over the domain using the unit eigenvector
+    /// let [x_name, y_name] = ufs.xy_fields::<KOLShapeFn>("unit_fields", unit_solution).unwrap();
+    ///
+    /// assert_eq!(x_name, String::from("unit_fields_x"));
+    /// assert_eq!(y_name, String::from("unit_fields_y"));
+    /// ```
     pub fn xy_fields<SF: ShapeFn>(
         &mut self,
         vector_name: &'static str,
-        eigenvector: Vec<f64>,
+        solution: Vec<f64>,
     ) -> Result<[String; 2], String> {
-        if eigenvector.len() != self.domain.dofs.len() {
+        if solution.len() != self.domain.dofs.len() {
             Err(format!(
[…]
