+++ src/integration/integrals.rs
@@ -1,9 +1,7 @@
-pub mod glq;
+use super::glq::*;
 use super::{Integral, IntegralResult};
 use crate::basis::{BasisFn, ShapeFn};
 use crate::domain::{dof::basis_spec::BasisDir, mesh::element::Materials, mesh::space::V2D};
-
-use glq::*;
 
 /// <∇ × u, ∇ × ρ>
 pub mod curl_curl {

[…]
