### 2022-04-08 227338f
made hp-refinement error handling more straightforward. Improved documentation on Mesh methods; specifically pertaining to error-handling
