+++ src/domain/mesh.rs
@@ -509,11 +509,13 @@ impl Mesh {
     /// mesh.h_refine_elems(vec![1], HRef::T).unwrap();
     /// mesh.h_refine_elems(vec![5], HRef::U(None)).unwrap();
     ///
-    /// let ancestors_with_start = mesh.ancestor_elems(10, true).unwrap();
-    /// assert_eq!(ancestors_with_start, vec![10, 5, 1, 0]);
+    /// let ancestors_with_start: [usize; 4] =
+    ///     mesh.ancestor_elems(10, true).unwrap().try_into().unwrap();
+    /// assert_eq!(ancestors_with_start, [10, 5, 1, 0]);
     ///
-    /// let ancestors_without_start = mesh.ancestor_elems(10, false).collect();
-    /// assert_eq!(ancestors_without_start, vec![5, 1, 0]);
+    /// let ancestors_without_start: [usize; 3] =
+    ///     mesh.ancestor_elems(10, false).unwrap().try_into().unwrap();
+    /// assert_eq!(ancestors_without_start, [5, 1, 0]);
     /// ```
     pub fn ancestor_elems(
         &self,
@@ -734,6 +736,7 @@ impl Mesh {
     /// mesh.h_refine_elems(vec![2, 3, 4], HRef::V(None)).unwrap();
     /// assert_eq!(mesh.elems.len(), 11);
     ///
+    /// assert!(mesh.h_refine_elems(vec![15], HRef::T).is_err());
     /// assert!(mesh.h_refine_elems(vec![0], HRef::T).is_err());
     /// assert!(mesh.h_refine_elems(vec![1, 1], HRef::T).is_err());
     /// ```
@@ -741,7 +744,7 @@ impl Mesh {
         if elems
             .iter()
             .enumerate()
-            .any(|(i, eid)| elems.iter().skip(i).any(|eid_t| eid_t == eid))
+            .any(|(i, eid)| elems.iter().skip(i + 1).any(|eid_t| eid_t == eid))
         {
             return Err(HRefError::DuplicateElemIds);
         }
@@ -770,7 +773,7 @@ impl Mesh {
     ///     } else {
     ///         None
     ///     }
-    /// }).unwrap();
+    /// });
     /// assert_eq!(mesh.elems.len(), 9);
     /// ```
     pub fn h_refine_with_filter<F>(&mut self, filt: F)
@@ -1209,11 +1212,11 @@ impl Mesh {
     /// use fem_2d::prelude::*;
     ///
     /// let mut mesh = Mesh::unit();
-    /// mesh.global_h_refinement(HRef::T).unwrap();
-    /// mesh.global_h_refinement(HRef::T).unwrap();
+    /// mesh.global_h_refinement(HRef::T);
+    /// mesh.global_h_refinement(HRef::T);
     ///
     /// // globally refine the mesh with [+3, +5] (where expansion orders start at [1, 1])
-    /// mesh.global_p_refinement(PRef::from(3, 5)).unwrap();
+    /// mesh.global_p_refinement(PRef::from(3, 5));
     /// for elem in mesh.elems.iter() {
     ///     assert_eq!(elem.poly_orders.ni, 4);
     ///     assert_eq!(elem.poly_orders.nj, 6);
@@ -1223,9 +1226,10 @@ impl Mesh {
     /// positive and negative p-refinements are constrained to fit within the valid range on each [Elem]
     /// ```
     /// use fem_2d::prelude::*;
+    /// use fem_2d::domain::mesh::MAX_POLYNOMIAL_ORDER;
     ///
     /// let mut mesh = Mesh::unit();
-    /// mesh.global_h_refinement(HRef::T).unwrap();
[…]
