### 2022-04-11 aa074a1
added error type for nalgebra-eigensolver. Better documentation for BasisFnSampler
