+++ src/basis.rs
@@ -25,7 +25,13 @@ pub trait ShapeFn: Clone + Sync + Send + std::fmt::Debug {
     fn poly_d2(&self, n: usize, p: usize) -> f64;
 }
 
-/// Structure used to generate and cache [BasisFn]'s over [Elem]'s in a Domain.
+/// Structure used to generate and cache [BasisFn]'s over [Elem]'s in a Domain
+///
+/// Three important areas of functionality are handled here:
+/// * Convert `BasisSpec`s (abstract basis function definitions on an `Elem`) into [BasisFn]'s (numerical instantiations of basis functions for integration)
+/// * Handle interpolation between parametric spaces to set up inter-layer integration
+/// * Cache computed [BasisFn]'s to avoid re-computation.
+///
 #[derive(Debug)]
 pub struct BasisFnSampler<SF: ShapeFn> {
     /// Maximum u-directed expansion order. [BasisFn]s Generated by this sampler will be defined up to this order in `i`
