### 2022-04-11 072c368
added error type for nalgebra-eigensolver. Better documentation for BasisFnSampler
