+++ src/basis.rs
@@ -25,10 +25,10 @@ pub trait ShapeFn: Clone + Sync + Send + std::fmt::Debug {
     fn poly_d2(&self, n: usize, p: usize) -> f64;
 }
 
-/// Structure used to generate and cache [BasisFn]'s over [Elem]'s in a Domain
+/// A utility structure used to generate and cache [BasisFn]'s during Integration
 ///
 /// Three important areas of functionality are handled here:
-/// * Convert `BasisSpec`s (abstract basis function definitions on an `Elem`) into [BasisFn]'s (numerical instantiations of basis functions for integration)
+/// * Convert `BasisSpec`s (abstract basis function definitions on an [Elem]) into [BasisFn]'s (numerical instantiations of basis functions for integration)
 /// * Handle interpolation between parametric spaces to set up inter-layer integration
 /// * Cache computed [BasisFn]'s to avoid re-computation.
 ///
@@ -56,6 +56,12 @@ impl<SF: ShapeFn> BasisFnSampler<SF> {
     /// * `num_u_points` : number of Gauss-Leg-Quad points to define along the u-axis. If `None`: a default value is used
     /// * `num_v_points` : number of Gauss-Leg-Quad points to define along the v-axis. If `None`: a default value is used
     /// * `compute_2nd_derivs`: whether or not to compute the second derivatives of the [ShapeFn]s.
+    ///
+    /// A [ShapeFn] implementation must also be specified.
+    ///
+    /// These parameters will be used to construct all [BasisFn]s generated by this sampler.
+    ///
+    /// The default values for `num_u_points` and `num_v_points` are cautiously large. Smaller values will result in much faster computation, but may result in a loss of accuracy depending on the implementation of [BasisFn]. For `KOLBasis`, n+2 is sufficient for accurate integration.
     pub fn with(
         i_max: usize,
         j_max: usize,
@@ -86,6 +92,8 @@ impl<SF: ShapeFn> BasisFnSampler<SF> {
     }
 
     /// Generate or retrieve a [BasisFn] defined over an [Elem]. Can be defined over a subset of the Element.
+    ///
+    /// If a descendant `Elem` is provided the [BasisFn] will be sampled over the region of the `Elem` that is overlapped by the descendant. Otherwise, the [BasisFn] will be sampled over the entire `Elem`.
     pub fn sample_basis_fn(
         &mut self,
         elem: &Elem,
