### 2022-04-11 fcf3d89
Improved error handling (#8)
* made hp-refinement error handling more straightforward. Improved documentation on Mesh methods; specifically pertaining to error-handling

* all doc tests passing

* slightly updated Domain docs

* added error type for nalgebra-eigensolver. Better documentation for BasisFnSampler

* added error type for nalgebra-eigensolver. Better documentation for BasisFnSampler

* Better docs and error handling on UniformFieldSpace
