+++ src/domain/fields.rs
@@ -2,6 +2,8 @@ use super::{dof::basis_spec::BasisDir, mesh::space::V2D, Domain};
 use crate::basis::{BasisFn, ShapeFn};
 
 use std::collections::{BTreeMap, HashMap};
+use std::error::Error;
+use std::fmt;
 use std::fs::File;
 use std::io::{BufWriter, Write};
 use std::time::SystemTime;
@@ -41,7 +43,8 @@ impl<'d> UniformFieldSpace<'d> {
     /// Use an eigenvector and associated [ShapeFn] to compute the X and Y fields over the [Domain]
     ///
     /// The X and Y field quantities will be stored as {vector_name}_x and {vector_name}_y respectively. The Names are returned in an array in that order.
-    ///
+    ///     
+    /// # Example
     /// ```
     /// use fem_2d::prelude::*;
     ///
@@ -61,12 +64,11 @@ impl<'d> UniformFieldSpace<'d> {
         &mut self,
         vector_name: &'static str,
         solution: Vec<f64>,
-    ) -> Result<[String; 2], String> {
+    ) -> Result<[String; 2], UniformFieldError> {
         if solution.len() != self.domain.dofs.len() {
-            Err(format!(
-                "NDofs != Solution length ({} != {}); Cannot compute xy fields over Domain",
+            Err(UniformFieldError::MismatchedSolutionSize(
                 self.domain.dofs.len(),
-                solution.len()
+                solution.len(),
             ))
         } else {
             let x_q_name = format!("{}_x", vector_name);
@@ -129,7 +131,9 @@ impl<'d> UniformFieldSpace<'d> {
     /// create a VTK file at the designated `path` (with the file `name.vtk`) including all Field Quantities
     ///
     /// These files can be plotted using [Visit](https://wci.llnl.gov/simulation/computer-codes/visit)
-    pub fn print_all_to_vtk(&self, path: impl AsRef<str>) -> std::io::Result<()> {
+    ///
+    /// Can return an IO error if the file cannot be written
+    pub fn print_all_to_vtk(&self, path: impl AsRef<str>) -> Result<(), Box<dyn Error>> {
         let all_q_names = self.quantities.keys().cloned().collect();
         self.print_quantities_to_vkt(path, all_q_names)
     }
@@ -137,11 +141,19 @@ impl<'d> UniformFieldSpace<'d> {
     /// create a VTK file at the designated `path` (with the file `name.vtk`) including a list of Field Quantities
     ///
     /// These files can be plotted using [Visit](https://wci.llnl.gov/simulation/computer-codes/visit)
+    ///
+    /// Can return an IO error if the file cannot be written, or a `UniformFieldError` if any of the quantity names are not found in the Field Space
     pub fn print_quantities_to_vkt(
         &self,
         path: impl AsRef<str>,
         quantity_names: Vec<String>,
-    ) -> std::io::Result<()> {
+    ) -> Result<(), Box<dyn Error>> {
+        for qn in quantity_names.iter() {
+            if !self.quantities.contains_key(qn) {
+                return Err(UniformFieldError::MissingQuantity(qn.clone()))?;
+            }
+        }
+
         let output_file = File::create(path.as_ref())?;
         let mut writer = BufWriter::new(&output_file);
 
@@ -215,23 +227,19 @@ impl<'d> UniformFieldSpace<'d> {
         // field values
         writeln!(writer, "POINT_DATA {}", num_points)?;
         for q_name in quantity_names {
-            match self.quantities.get(&q_name) {
-                Some(field_quant) => {
-                    field_quant.write_vtk_quantity(&mut writer)?;
-                }
-                None => println!(
[…]
