+++ src/domain.rs
@@ -29,7 +29,7 @@ pub struct Domain {
 }
 
 impl Domain {
-    /// Create a Domain around a Blank-Mesh
+    /// Create a Domain around an empty mesh
     pub fn blank() -> Self {
         Self {
             mesh: Mesh::blank(),
@@ -38,7 +38,7 @@ impl Domain {
         }
     }
 
-    /// Create a Domain from a Unit-Mesh
+    /// Create a Domain from a unit Mesh
     pub fn unit() -> Self {
         Self::from_mesh(Mesh::unit())
     }
@@ -72,7 +72,11 @@ impl Domain {
 
     /// Retrieve a [BasisSpec] at a particular [BSAddress]
     ///
-    /// Returns an error if the designated `Elem` does not exist, or does not have that [BasisSpec]
+    /// # Returns
+    /// * A reference to the [BasisSpec] if it exists
+    /// * An `Err` if the `elem` does not exist
+    /// * An `Err` if the `elem` does not have a [BasisSpec] at the given `address`
+    ///
     pub fn get_basis_spec(&self, bs_address: BSAddress) -> Result<&BasisSpec, String> {
         if bs_address.elem_id > self.mesh.elems.len() {
             Err(format!(
@@ -205,6 +209,7 @@ impl Domain {
 
     /// Retrieve a list of [BasisSpec]s on an `Elem` by ID
     ///
+    /// # Example
     /// ```
     /// use fem_2d::prelude::*;
     /// let mut mesh = Mesh::unit();
@@ -218,12 +223,9 @@ impl Domain {
     /// // 2 elem-type basis specs in each direction
     /// assert_eq!(basis_specs.len(), 4);
     /// ```
-    pub fn local_basis_specs(&self, elem_id: usize) -> Result<&Vec<BasisSpec>, String> {
+    pub fn local_basis_specs(&self, elem_id: usize) -> Result<&Vec<BasisSpec>, MeshAccessError> {
         if elem_id >= self.mesh.elems.len() {
-            Err(format!(
-                "Elem {} doesn't exist; Cannot retrieve BasisSpecs!",
-                elem_id
-            ))
+            Err(MeshAccessError::ElemDoesNotExist(elem_id))
         } else {
             Ok(&self.basis_specs[elem_id])
         }
@@ -231,21 +233,26 @@ impl Domain {
 
     /// Retrieve a list of an `Elem`s descendant [BasisSpec]s (All the [`BasisSpec`]s on its descendant `Elem`s)
     ///
-    /// Lists are returned as tuples of the form `(elem_id, [basis_specs])`
+    /// # Returns
+    /// * A vector of tuples in the form: `(desc_elem_id: usize, desc_elem_basis_specs: Vec<BasisSpec>)`. Where each tuple corresponds to one descendant `Elem`
+    /// * An error if the given `elem_id` does not exist
     ///
+    /// The [BasisSpec]s on `elem_id` itself are not included in the returned vector.
+    ///
+    /// # Example
     /// ```
     /// use fem_2d::prelude::*;
     ///
     /// let mut mesh = Mesh::unit();
     /// mesh.set_global_expansion_orders([2, 2]).unwrap();
-    /// mesh.global_h_refinement(HRef::T).unwrap();
+    /// mesh.global_h_refinement(HRef::T);
     ///
     /// let dom = Domain::from_mesh(mesh);
     ///
     /// // get the basis specs from all of Elem 0's descendants
     /// let basis_specs = dom.descendant_basis_specs(0).unwrap();
     ///
-    /// // Elem 0 has 4 descendant Elems due to the T-Type h-Refinement
[…]
