### 2022-04-19 e3ca469
merging error-handling changes
