+++ src/basis.rs
@@ -326,14 +326,17 @@ impl<SF: ShapeFn> BasisFn<SF> {
         }
     }
 
+    /// Evaluate the u-directed basis function at some point (m, n)
     pub fn f_u(&self, [i, j]: [usize; 2], [m, n]: [usize; 2]) -> V2D {
         self.ti[m][n].u * self.u_shapes.power(i, m) * self.v_shapes.poly(j, n)
     }
 
+    /// Evaluate the v-directed basis function at some point (m, n)
     pub fn f_v(&self, [i, j]: [usize; 2], [m, n]: [usize; 2]) -> V2D {
         self.ti[m][n].v * self.u_shapes.poly(i, m) * self.v_shapes.power(j, n)
     }
 
+    /// Evaluate the first derivative of the u-directed basis with respect to another `Elem`'s parametric space
     pub fn f_u_d1(&self, [i, j]: [usize; 2], [m, n]: [usize; 2], para_scale: &V2D) -> V2D {
         self.ti[m][n].u
             * V2D::from([
@@ -343,6 +346,7 @@ impl<SF: ShapeFn> BasisFn<SF> {
             * para_scale
     }
 
+    /// Evaluate the first derivative of the v-directed basis with respect to another `Elem`'s parametric space
     pub fn f_v_d1(&self, [i, j]: [usize; 2], [m, n]: [usize; 2], para_scale: &V2D) -> V2D {
         self.ti[m][n].v
             * V2D::from([
@@ -352,6 +356,7 @@ impl<SF: ShapeFn> BasisFn<SF> {
             * para_scale
     }
 
+    /// Evaluate the second derivative of the u-directed basis with respect to another `Elem`'s parametric space
     pub fn f_u_d2(&self, [i, j]: [usize; 2], [m, n]: [usize; 2], para_scale: &V2D) -> V2D {
         self.ti[m][n].u
             * V2D::from([
@@ -362,6 +367,7 @@ impl<SF: ShapeFn> BasisFn<SF> {
             * para_scale
     }
 
+    /// Evaluate the second derivative of the v-directed basis with respect to another `Elem`'s parametric space
     pub fn f_v_d2(&self, [i, j]: [usize; 2], [m, n]: [usize; 2], para_scale: &V2D) -> V2D {
         self.ti[m][n].v
             * V2D::from([
@@ -372,6 +378,7 @@ impl<SF: ShapeFn> BasisFn<SF> {
             * para_scale
     }
 
+    /// Evaluate the gradient of the u-directed basis with respect to another `Elem`'s parametric space
     pub fn f_u_dd(&self, [i, j]: [usize; 2], [m, n]: [usize; 2], para_scale: &V2D) -> V2D {
         self.ti[m][n].u
             * self.u_shapes.power_d1(i, m)
@@ -380,6 +387,7 @@ impl<SF: ShapeFn> BasisFn<SF> {
             * para_scale[1]
     }
 
+    /// Evaluate the gradient of the v-directed basis with respect to another `Elem`'s parametric space
     pub fn f_v_dd(&self, [i, j]: [usize; 2], [m, n]: [usize; 2], para_scale: &V2D) -> V2D {
         self.ti[m][n].v
             * self.u_shapes.poly_d1(i, m)
@@ -389,11 +397,13 @@ impl<SF: ShapeFn> BasisFn<SF> {
     }
 
     #[inline]
+    /// The size of the parametric area relative to the unit-parametric area
     pub fn glq_scale(&self) -> f64 {
         self.para_scale[0] * self.para_scale[1]
     }
 
     #[inline]
+    /// The size of the parametric space relative to the unit-parametric space (along a single edge)
[…]
