### 2022-04-20 8e39995
modes regrouped into two parent modules. Improved ShapeFn and BasisFn documentation and method names. Improved module documentation in Domain and Mesh.
