+++ src/matrix_math.rs
@@ -1,0 +1,4 @@
+/// Structures and functions to assist in the integration of Basis Functions
+pub mod integration;
+/// Structures and functions to solve Generalized Eigenvalue Problems
+pub mod linalg;
