### 2022-04-20 d8ea44b
improved BasisFn documentation
