### 2022-04-25 3575aef
removed galerking sampling functionality from Domain. Samping is in its own module s.t. new types of samplers can be implemented easily with their own fn headers (and trait bounds).
