+++ src/fem_domain/domain.rs
@@ -5,19 +5,13 @@ pub mod fields;
 /// The internal geometric structure of a Domain. This is modified by hp-refinements.
 pub mod mesh;
 
-use super::basis::{BasisFnSampler, ParBasisFnSampler, ShapeFn};
-use crate::matrix_math::integration::Integral;
-use crate::matrix_math::linalg::{sparse_matrix::SparseMatrix, GEP};
 use dof::{
     basis_spec::{BSAddress, BasisDir, BasisLoc, BasisSpec},
     DoF,
 };
 use mesh::*;
-
-use rayon::prelude::*;
 use smallvec::smallvec;
 use std::collections::BTreeMap;
-use std::sync::{Arc, Mutex};
 
 /// High Level Description of an FEM Domain
 pub struct Domain {
@@ -336,292 +330,292 @@ impl Domain {
         BSAddress::new(elem_id, elem_idx)
     }
 
-    /// Fill two system matrices using the domains it's Basis Space as the Testing Space. Return a Generalized Eigenproblem ([GEP])
-    ///
-    /// All pairs of overlapping Shape Functions will be integrated and stored in the matrices by their associated DoF IDs
-    ///
-    /// * Two [Integral]s: `AI` and `BI` must be specified. These are used to populate the A and B matrices respectively
-    /// * A [ShapeFn] `SF` must also be specified. This is used to evaluate the Domains [BasisSpec]s
-    pub fn galerkin_sample_gep<SF, AI, BI>(&self, num_gauss_quad: Option<usize>) -> GEP
-    where
-        SF: ShapeFn,
-        AI: Integral,
-        BI: Integral,
-    {
-        // construct an eigenproblem with a and b matrices
-        let mut gep = GEP::new(self.dofs.len());
+    // /// Fill two system matrices using the domains it's Basis Space as the Testing Space. Return a Generalized Eigenproblem ([GEP])
+    // ///
+    // /// All pairs of overlapping Shape Functions will be integrated and stored in the matrices by their associated DoF IDs
+    // ///
+    // /// * Two [Integral]s: `AI` and `BI` must be specified. These are used to populate the A and B matrices respectively
+    // /// * A [ShapeFn] `SF` must also be specified. This is used to evaluate the Domains [BasisSpec]s
+    // pub fn galerkin_sample_gep<SF, AI, BI>(&self, num_gauss_quad: Option<usize>) -> GEP
+    // where
+    //     SF: ShapeFn,
+    //     AI: Integral,
+    //     BI: Integral,
+    // {
+    //     // construct an eigenproblem with a and b matrices
+    //     let mut gep = GEP::new(self.dofs.len());
 
-        // construct basis sampler
-        let [i_max, j_max] = self.mesh.max_expansion_orders();
-        let (mut bf_sampler, [u_weights, v_weights]): (BasisFnSampler<SF>, _) =
-            BasisFnSampler::with(
-                i_max as usize,
-                j_max as usize,
-                num_gauss_quad,
-                num_gauss_quad,
-                false,
-            );
+    //     // construct basis sampler
+    //     let [i_max, j_max] = self.mesh.max_expansion_orders();
+    //     let (mut bf_sampler, [u_weights, v_weights]): (BasisFnSampler<SF>, _) =
+    //         BasisFnSampler::with(
+    //             i_max as usize,
+    //             j_max as usize,
+    //             num_gauss_quad,
+    //             num_gauss_quad,
+    //             false,
+    //         );
 
-        // setup integration
-        let a_integrator = AI::with_weights(&u_weights, &v_weights);
-        let b_integrator = BI::with_weights(&u_weights, &v_weights);
+    //     // setup integration
[…]
