### 2022-04-25 dad92d6
working on improving names of shape function and basis function types. Will set up a hierarchy of basis fn traits to make implementation of alteranate types easier (ex: div-conforming).
