+++ src/fem_domain/basis.rs
@@ -1,5 +1,5 @@
-/// Specific implementations of the `ShapeFn` Trait (MaxOrthoShapeFn can be added as a Feature on the Nightly Toolchain)
-pub mod shape_fns;
+/// Specific implementations of the `HierBasisFnSpace` Trait (MaxOrthoShapeFn can be added as a Feature on the Nightly Toolchain)
+pub mod hierarchical_basis_fns;
 
 use super::domain::mesh::{
     elem::Elem,
@@ -10,18 +10,18 @@ use crate::matrix_math::integration::glq::{gauss_quadrature_points, scale_gauss_
 use std::collections::HashMap;
 use std::sync::{Arc, Mutex};
 
-/// A 1D Higher-Order Hierarchical Shape Function Space: defined on (-1.0, +1.0)
+/// One Directional Component of a Higher-Order Vectorial Hierarchical Basis Function Space: defined on (-1.0, +1.0)
 ///
-/// Outer products of the 1D vectors in this space are used to define 2D Basis Functions via the [BasisFn] struct
+/// Outer products of the 1D vectors in pairs of HierBasisFnSpaces are used to define 2D Basis Functions ([HierBasisFnSampled]) via the [BasisFnSampler]
 ///
 /// Two sets of shape vectors must be defined: Tangential and Normal
 /// * The tangential shapes are sampled along the axis that is tangential to the direction of the basis function
 /// * The normal shapes are sampled along the axis that is normal to the direction of the basis function
 ///
-/// For example: a u-directed basis function with orders (2, 3) will sample `shape.tang(2, m) * shape.norm(3, n)`, and a v-directed basis function with orders (4, 1) will sample `shape.norm(4, m) * shape.tang(1, n)` for all points `m` and `n`
+/// For example: a u-directed basis function with orders (2, 3) will sample `u_shape.tang(2, m) * v_shape.norm(3, n)`, and a v-directed basis function with orders (4, 1) will sample `u_shape.norm(4, m) * v_shape.tang(1, n)` for all points `m` and `n`
 ///
 /// We also define the 1st derivative, and optionally the 2nd derivative of the shape functions in order to compute gradients of the Basis Functions.
-pub trait ShapeFn: Clone + Sync + Send + std::fmt::Debug {
+pub trait HierBasisFnSpace: Clone + Sync + Send + std::fmt::Debug {
     /// Define a set of shapes (or functions) over a set of points
     ///
     /// # Arguments
@@ -46,100 +46,6 @@ pub trait ShapeFn: Clone + Sync + Send + std::fmt::Debug {
     fn norm_d2(&self, n: usize, p: usize) -> f64;
 }
 
-// /// A utility structure used to generate and cache [BasisFn]'s during Integration
-// ///
-// /// Three important areas of functionality are handled here:
-// /// * Convert `BasisSpec`s (abstract basis function definitions on an [Elem]) into [BasisFn]'s (numerical instantiations of basis functions for integration)
-// /// * Handle interpolation between parametric spaces to set up inter-layer integration
-// /// * Cache computed [BasisFn]'s to avoid re-computation.
-// ///
-// #[derive(Debug)]
-// pub struct BasisFnSampler<SF: ShapeFn> {
-//     /// Maximum u-directed expansion order. [BasisFn]s Generated by this sampler will be defined up to this order in `i`
-//     pub i_max: usize,
-//     /// Maximum v-directed expansion order. [BasisFn]s Generated by this sampler will be defined up to this order in `i`
-//     pub j_max: usize,
-//     /// Whether the 2nd derivatives of the [ShapeFn]s will be computed. If true, the endpoints [-1, +1] are also defined along both axes!
[…]
