### 2022-04-25 1fe5421
BasisFnSampler is generic over any type of BasisFn. Still working on renaming BasisFn related structures/traits, and improving documentation
