+++ src/fem_domain/basis.rs
@@ -10,17 +10,12 @@ use crate::matrix_math::integration::glq::{gauss_quadrature_points, scale_gauss_
 use std::collections::HashMap;
 use std::sync::{Arc, Mutex};
 
-/// One Directional Component of a Higher-Order Vectorial Hierarchical Basis Function Space: defined on (-1.0, +1.0)
+/// A Trait to define the functional space used to compose a [HierBasisFnSampled]
 ///
-/// Outer products of the 1D vectors in pairs of HierBasisFnSpaces are used to define 2D Basis Functions ([HierBasisFnSampled]) via the [BasisFnSampler]
+/// This Trait requires that the T (tangential) and N (normal) vector spaces be defined over the given set of `points` (which fall within the range [-1, 1]) up to the given expansion order.
 ///
-/// Two sets of shape vectors must be defined: Tangential and Normal
-/// * The tangential shapes are sampled along the axis that is tangential to the direction of the basis function
-/// * The normal shapes are sampled along the axis that is normal to the direction of the basis function
+/// The first (and optionally the second) derivative of the spaces must also be defined over the same set of points.
 ///
-/// For example: a u-directed basis function with orders (2, 3) will sample `u_shape.tang(2, m) * v_shape.norm(3, n)`, and a v-directed basis function with orders (4, 1) will sample `u_shape.norm(4, m) * v_shape.tang(1, n)` for all points `m` and `n`
-///
-/// We also define the 1st derivative, and optionally the 2nd derivative of the shape functions in order to compute gradients of the Basis Functions.
 pub trait HierBasisFnSpace: Clone + Sync + Send + std::fmt::Debug {
     /// Define a set of shapes (or functions) over a set of points
     ///
@@ -29,21 +24,23 @@ pub trait HierBasisFnSpace: Clone + Sync + Send + std::fmt::Debug {
     /// * `points` - The set of points to evaluate the shape functions over
     /// * `compute_d2` - Whether to compute the second derivative of the shape functions
     ///
+    /// If `compute_d2` is `false`, then calls to `norm_d2` and `tang_d2` should not be expected
+    ///
     fn with(max_order: usize, points: &[f64], compute_d2: bool) -> Self;
 
     /// Sample the `n`th order tangentially orientated shape function at the given point `p`
-    fn tang(&self, n: usize, p: usize) -> f64;
-    /// Sample the 1st derivative of the `n`th order tangentially orientated shape function at the given point `p`
-    fn tang_d1(&self, n: usize, p: usize) -> f64;
-    /// Sample the 2nd derivative of the `n`th order tangentially orientated shape function at the given point `p`. This method does not have to return if `compute_d2` was set to `false`.
-    fn tang_d2(&self, n: usize, p: usize) -> f64;
-
-    /// Sample the `n`th order normally orientated shape function at the given point `p`
     fn norm(&self, n: usize, p: usize) -> f64;
     /// Sample the 1st derivative of the `n`th order tangentially orientated shape function at the given point `p`
     fn norm_d1(&self, n: usize, p: usize) -> f64;
-    /// Sample the 2nd derivative of the `n`th order normally orientated shape function at the given point `p`. This method does not have to return if `compute_d2` was set to `false`.
[…]
