### 2022-04-25 5f73764
renamed some basis function structs
