+++ src/matrix_math/integration/integrals.rs
@@ -1,6 +1,6 @@
 use super::glq::*;
-use super::{Integral, IntegralResult};
-use crate::fem_domain::basis::{HierBasisFnSampled, HierBasisFnSpace};
+use super::{HierCurlIntegral, IntegralResult};
+use crate::fem_domain::basis::{HierCurlBasisFn, HierCurlBasisFnSpace};
 use crate::fem_domain::domain::{
     dof::basis_spec::BasisDir, mesh::element::Materials, mesh::space::V2D,
 };
@@ -15,7 +15,7 @@ pub mod curl_curl {
         v_weights: Vec<f64>,
     }
 
-    impl Integral for CurlCurl {
+    impl HierCurlIntegral for CurlCurl {
         fn with_weights(u_weights: &[f64], v_weights: &[f64]) -> Self {
             Self {
                 u_weights: u_weights.to_vec(),
@@ -23,14 +23,14 @@ pub mod curl_curl {
             }
         }
 
-        fn integrate<BSpace: HierBasisFnSpace>(
+        fn integrate<BSpace: HierCurlBasisFnSpace>(
             &self,
             p_dir: BasisDir,
             q_dir: BasisDir,
             p_orders: [usize; 2],
             q_orders: [usize; 2],
-            p_basis: &HierBasisFnSampled<BSpace>,
-            q_basis: &HierBasisFnSampled<BSpace>,
+            p_basis: &HierCurlBasisFn<BSpace>,
+            q_basis: &HierCurlBasisFn<BSpace>,
             materials: &Materials,
         ) -> IntegralResult {
             IntegralResult::Full(
@@ -91,14 +91,14 @@ pub mod curl_curl {
             )
         }
 
-        fn integrate_by_parts<BSpace: HierBasisFnSpace>(
+        fn integrate_by_parts<BSpace: HierCurlBasisFnSpace>(
             &self,
             p_dir: BasisDir,
             q_dir: BasisDir,
             p_orders: [usize; 2],
             q_orders: [usize; 2],
-            p_basis: &HierBasisFnSampled<BSpace>,
-            q_basis: &HierBasisFnSampled<BSpace>,
+            p_basis: &HierCurlBasisFn<BSpace>,
+            q_basis: &HierCurlBasisFn<BSpace>,
             materials: &Materials,
         ) -> IntegralResult {
             let surface_term = (1.0 / materials.mu_rel.re)
@@ -247,9 +247,9 @@ pub mod curl_curl {
     ];
 
     #[inline]
-    fn max_uv_ratios<BSpace: HierBasisFnSpace>(
-        p_basis: &HierBasisFnSampled<BSpace>,
-        q_basis: &HierBasisFnSampled<BSpace>,
+    fn max_uv_ratios<BSpace: HierCurlBasisFnSpace>(
+        p_basis: &HierCurlBasisFn<BSpace>,
+        q_basis: &HierCurlBasisFn<BSpace>,
         [m, n]: [usize; 2],
     ) -> f64 {
         ((p_basis.det_jac[m][n] >= q_basis.det_jac[m][n]) as u8) as f64 * p_basis.uv_ratio([m, n])
@@ -259,9 +259,9 @@ pub mod curl_curl {
     }
 
     #[inline]
-    fn max_vu_ratios<BSpace: HierBasisFnSpace>(
-        p_basis: &HierBasisFnSampled<BSpace>,
-        q_basis: &HierBasisFnSampled<BSpace>,
+    fn max_vu_ratios<BSpace: HierCurlBasisFnSpace>(
+        p_basis: &HierCurlBasisFn<BSpace>,
+        q_basis: &HierCurlBasisFn<BSpace>,
         [m, n]: [usize; 2],
     ) -> f64 {
[…]

[…]
