+++ src/fem_domain/domain.rs
@@ -329,293 +329,6 @@ impl Domain {
 
         BSAddress::new(elem_id, elem_idx)
     }
-
-    // /// Fill two system matrices using the domains it's Basis Space as the Testing Space. Return a Generalized Eigenproblem ([GEP])
-    // ///
-    // /// All pairs of overlapping Shape Functions will be integrated and stored in the matrices by their associated DoF IDs
-    // ///
-    // /// * Two [Integral]s: `AI` and `BI` must be specified. These are used to populate the A and B matrices respectively
-    // /// * A [ShapeFn] `SF` must also be specified. This is used to evaluate the Domains [BasisSpec]s
-    // pub fn galerkin_sample_gep<SF, AI, BI>(&self, num_gauss_quad: Option<usize>) -> GEP
-    // where
-    //     SF: ShapeFn,
-    //     AI: Integral,
-    //     BI: Integral,
-    // {
-    //     // construct an eigenproblem with a and b matrices
-    //     let mut gep = GEP::new(self.dofs.len());
-
-    //     // construct basis sampler
-    //     let [i_max, j_max] = self.mesh.max_expansion_orders();
-    //     let (mut bf_sampler, [u_weights, v_weights]): (BasisFnSampler<SF>, _) =
-    //         BasisFnSampler::with(
-    //             i_max as usize,
-    //             j_max as usize,
-    //             num_gauss_quad,
-    //             num_gauss_quad,
-    //             false,
-    //         );
-
-    //     // setup integration
-    //     let a_integrator = AI::with_weights(&u_weights, &v_weights);
-    //     let b_integrator = BI::with_weights(&u_weights, &v_weights);
-
-    //     for elem in self.elems() {
-    //         // get relevant data for this Elem
-    //         let local_basis_specs = self.local_basis_specs(elem.id).unwrap();
-    //         let desc_basis_specs = self.descendant_basis_specs(elem.id).unwrap();
-
-    //         let bs_local = bf_sampler.sample_basis_fn(elem, None);
-    //         let elem_materials = elem.get_materials();
-
-    //         let mut local_a_entries: Vec<([usize; 2], f64)> =
-    //             Vec::with_capacity(local_basis_specs.len() * local_basis_specs.len() / 2);
-    //         let mut local_b_entries: Vec<([usize; 2], f64)> =
-    //             Vec::with_capacity(local_basis_specs.len() * local_basis_specs.len() / 2);
-
-    //         // local - local
-    //         for (i, (p_orders, p_dir, p_dof_id)) in local_basis_specs
-    //             .iter()
-    //             .map(|bs_p| bs_p.integration_data())
-    //             .enumerate()
-    //         {
-    //             for (q_orders, q_dir, q_dof_id) in local_basis_specs
-    //                 .iter()
-    //                 .skip(i)
-    //                 .map(|bs_q| bs_q.integration_data())
-    //             {
-    //                 let a = a_integrator
-    //                     .integrate(
-    //                         p_dir,
-    //                         q_dir,
-    //                         p_orders,
-    //                         q_orders,
-    //                         &bs_local,
-    //                         &bs_local,
-    //                         &elem_materials,
-    //                     )
-    //                     .full_solution();
-    //                 let b = b_integrator
-    //                     .integrate(
-    //                         p_dir,
-    //                         q_dir,
-    //                         p_orders,
-    //                         q_orders,
-    //                         &bs_local,
-    //                         &bs_local,
-    //                         &elem_materials,
-    //                     )
-    //                     .full_solution();
-
-    //                 local_a_entries.push(([p_dof_id, q_dof_id], a));
-    //                 local_b_entries.push(([p_dof_id, q_dof_id], b));
-    //             }
-    //         }
-
-    //         gep.a.insert_group(local_a_entries);
[…]
