### 2022-04-26 d39cfa8
added explicit continuity condition option to Domain construction. This is checked by galerkin sampling function. Improved documentation on BasisSpec and galerkin sampling functions
