+++ src/fem_domain/domain.rs
@@ -12,41 +12,150 @@ use dof::{
 use mesh::*;
 use smallvec::smallvec;
 use std::collections::BTreeMap;
+use std::fmt;
+
+/// The Continuity Condition to be enforced by the Domain. Only H(Curl) is currently supported!!!
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum ContinuityCondition {
+    HCurl,
+    HDiv,
+    Discontinuous,
+}
+
+impl fmt::Display for ContinuityCondition {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        match self {
+            Self::HCurl => write!(f, "H(Curl)"),
+            Self::HDiv => write!(f, "H(Div)"),
+            Self::Discontinuous => write!(f, "Discontinuous"),
+        }
+    }
+}
 
 /// High Level Description of an FEM Domain
+///
+/// This struct contains:
+/// * The [Mesh] - A discretization of the Domain of interest
+/// * A list of [DoF]s (Degrees of Freedom) - Sets of Basis Functions that can take a scalar value in the FEM solution
+/// * A list of [BasisSpec]s - The specification of each Basis Function in the Domain
+///
 pub struct Domain {
     pub mesh: Mesh,
     /// Degrees of Freedom (collections of BasisSpecs matched via Tangential Continuity)
     pub dofs: Vec<DoF>,
     /// Individual BasisSpecs sorted by their associated Elem
     pub basis_specs: Vec<Vec<BasisSpec>>,
+    /// The continuity condition enforced over the Basis Space
+    pub cc: ContinuityCondition,
 }
 
 impl Domain {
     /// Create a Domain around an empty mesh
-    pub fn blank() -> Self {
+    pub fn blank(cc: ContinuityCondition) -> Self {
         Self {
             mesh: Mesh::blank(),
             dofs: Vec::new(),
             basis_specs: Vec::new(),
+            cc,
         }
     }
 
     /// Create a Domain from a unit Mesh
-    pub fn unit() -> Self {
-        Self::from_mesh(Mesh::unit())
+    pub fn unit(cc: ContinuityCondition) -> Self {
+        Self::from_mesh(Mesh::unit(), cc)
     }
 
     /// Create a Domain from a Mesh
-    pub fn from_mesh(mesh: Mesh) -> Self {
-        let mut dom = Self {
-            mesh,
-            dofs: Vec::new(),
-            basis_specs: Vec::new(),
-        };
+    pub fn from_mesh(mut mesh: Mesh, cc: ContinuityCondition) -> Self {
+        // prepare for basis function matching
+        mesh.set_edge_activation();
+        // mesh.set_node_activation(); (TODO: this is not needed until node-type basis functions are implemented)
 
-        dom.gen_dofs();
-        dom
+        // create dof and basis_spec collections
+        let mut dof_id_tracker = IdTracker::new(0);
+        let mut basis_specs = vec![Vec::new(); mesh.elems.len()];
+        let mut dofs = Vec::new();
+
+        // Generate lists of BasisSpecs associated with Elems, Edges, and (Nodes), sorted by their IDs
+        let [elem_bs, edge_bs, _] = Self::gen_basis_specs(&mesh, cc);
+
+        // Designate all elem-type BasisSpecs located on shell Elems as DoFs
+        for (elem_id, mut elem_bs_list) in elem_bs {
+            if !mesh.elems[elem_id].has_children() {
+                basis_specs[elem_id] = Vec::with_capacity(elem_bs_list.len());
+
+                for elem_bs in elem_bs_list
+                    .drain(0..)
+                    .filter(|bs| bs.dir == BasisDir::U || bs.dir == BasisDir::V)
[…]
