### 2022-04-27 74a9936
updated doc tests to reflect design changes and module structure changes
