### 2022-04-27 f7607d5
Basis fn design changes (#9)
* BasisFn documentation

* modes regrouped into two parent modules. Improved ShapeFn and BasisFn documentation and method names. Improved module documentation in Domain and Mesh.

* improved BasisFn documentation

* removed galerking sampling functionality from Domain. Samping is in its own module s.t. new types of samplers can be implemented easily with their own fn headers (and trait bounds).

* working on improving names of shape function and basis function types. Will set up a hierarchy of basis fn traits to make implementation of alteranate types easier (ex: d
[…]
