+++ src/fem_problem/galerkin.rs
@@ -1,0 +1,209 @@
+use super::{
+    integration::HierCurlIntegral,
+    linalg::{sparse_matrix::SparseMatrix, GEP},
+};
+use crate::fem_domain::{
+    basis::{BasisFnSampler, HierCurlBasisFn, HierCurlBasisFnSpace},
+    domain::{ContinuityCondition, Domain},
+};
+use rayon::prelude::*;
+use std::fmt;
+
+/// Minimum number of Gauss Legendre Quadrature Points Allowed for Galerkin Sampling
+pub const MIN_GLQ_ORDER: usize = 4;
+
+/// Fill two system matrices using a [Domain]'s Basis Space as the Testing Space. Return a Generalized Eigenproblem ([GEP])
+///
+/// All pairs of overlapping Shape Functions will be integrated and stored in the matrices by their associated DoF IDs
+///
+/// Computations are parallelized over the Rayon Global Threadpool
+///
+/// # Arguments
+/// * `domain`: The [Domain] over which the Galerkin Sampling is to be performed
+/// * `num_u_glq`: The number of Gauss Legendre Quadrature Points in to use for integration in the u-direction. If `None`, the default is used.
+/// * `num_v_glq`: The number of Gauss Legendre Quadrature Points in to use for integration in the v-direction. If `None`, the default is used.
+/// * Two [HierCurlIntegral]s: `AI` and `BI` must be specified as Generic Arguments. These are used to populate the A and B matrices respectively
+/// * A [HierCurlBasisFnSpace] `BSpace` must also be specified as a Generic Argument. This is used to instantiate the Domains `BasisSpec`s as [HierCurlBasisFn]s
+///
+/// # Returns
+/// * An `Err` if the `Domain` was not constructed with an `H(Curl)` [ContinuityCondition]
+/// * An `Err` if the `Domain` doesn't have any Degrees of Freedom
+/// * An `Err` if the specified number of Gauss Legendre Points is too small
+/// * A [GEP], otherwise
+///
+pub fn galerkin_sample_gep_hcurl<
+    BSpace: HierCurlBasisFnSpace,
+    AI: HierCurlIntegral,
+    BI: HierCurlIntegral,
+>(
+    domain: &Domain,
+    [num_u_glq, num_v_glq]: [Option<usize>; 2],
+) -> Result<GEP, GalerkinSamplingError> {
+    // check for errors
+    if domain.cc != ContinuityCondition::HCurl {
+        return Err(GalerkinSamplingError::WrongContinuityCondition(
+            ContinuityCondition::HCurl,
+            domain.cc,
+        ));
+    }
+    if domain.dofs.is_empty() {
+        return Err(GalerkinSamplingError::EmptyDOFSet);
+    }
+    if num_u_glq.lt(&Some(MIN_GLQ_ORDER)) || num_v_glq.lt(&Some(MIN_GLQ_ORDER)) {
+        return Err(GalerkinSamplingError::InvalidGLQSettings);
+    }
+
+    // construct an eigenproblem with a and b matrices
+    let mut gep = GEP::new(domain.dofs.len());
+
+    // construct basis sampler
+    let [i_max, j_max] = domain.mesh.max_expansion_orders();
+    let (bs_sampler, [u_weights, v_weights]): (BasisFnSampler<HierCurlBasisFn<BSpace>>, _) =
+        BasisFnSampler::with(i_max as usize, j_max as usize, num_u_glq, num_v_glq, false);
+
+    // setup integration
+    let a_integrator = AI::with_weights(&u_weights, &v_weights);
+    let b_integrator = BI::with_weights(&u_weights, &v_weights);
+
+    gep.par_extend(domain.mesh.elems.par_iter().map(|elem| {
[…]
