+++ src/fem_domain.rs
@@ -1,0 +1,4 @@
+/// Structures and Traits for Basis Function Evaluation
+pub mod basis;
+/// Structures to define the geometric characteristics and refinement state of an FEM Domain
+pub mod domain;
