+++ src/fem_problem.rs
@@ -1,0 +1,8 @@
+/// Structures and functions to assist in the integration of Basis Functions
+pub mod integration;
+
+/// Structures and functions to solve Generalized Eigenvalue Problems
+pub mod linalg;
+
+/// Strucutures to Execute Galerkin Sampling over a `Domain` using `Integral`s
+pub mod galerkin;
