### 2022-04-27 10422f0
separate slepc and nalgabra tests in lib
