### 2022-04-27 1be2c81
separate slepc and nalgabra tests in lib
