### 2022-04-27 ce06f4a
Create rust.yml
addresses #6
