+++ Cargo.toml
@@ -22,4 +22,4 @@ json = "0.12.4"
 nalgebra = "0.30.1"
 num-complex = "0.4.0"
 rayon = "1.5.1"
-smallvec = "1.7.0"
+smallvec = "1.8.0"
