+++ src/lib.rs
@@ -32,7 +32,7 @@ pub mod prelude {
     pub use crate::fem_problem::galerkin::galerkin_sample_gep_hcurl;
     pub use crate::fem_problem::integration::integrals::{curl_curl::CurlCurl, inner::L2Inner};
     pub use crate::fem_problem::linalg::{
-        nalgebra_solve::nalgebra_solve_gep,
+        nalgebra_solve::{nalgebra_solve_gep, NalgebraGEPError},
         slepc_solve::{slepc_solve_gep, SlepcGEPError},
         EigenPair, GEP,
     };
