### 2022-04-29 9335fe3
started extending paper to better explain the hp-refinement api (and overal project goals). Made some small documentation fixes
