+++ paper.md
@@ -25,64 +25,147 @@ date: 14 February 2022
 bibliography: paper.bib
 ---
 
-# Summary
+# Introduction
 
-The Finite Element Method (FEM) is a powerful tool used to solve Partial Differential Equations (PDE)s on arbitrary geometries. As its name suggests, the method works by breaking a geometric model (a Domain) into a set of small elements, each of which assigned a set of Basis Functions. A solution is expressed in terms of a weighted superposition of all the functions in the Domain such that the PDE is satisfied along with some continuity and boundary conditions. Some common applications include the Navier-Stokes equations which characterize the behavior of fluids, Schrödinger's equation which governs the evolution of quantum systems, and Maxwell's Equations which are a macroscopic description of essentially all Electromagnetic phenomena.
+The Finite Element Method (FEM) is a powerful computation methodology used to solve Partial Differential Equations (PDE)s on arbitrary geometries. In reality, physical systems behave in a spatiotemporally continuous manner; however FEM solvers are able to model these dynamics with a high fidelity by decomposing a physical model into a finite set of elements. Each element supports a finite number of degrees of freedom, which are used to describe the behavior of the system. 
 
-The `FEM_2D` library is a powerful FEM simulation tool implemented in Rust. It was designed within the domain of Computational Electromagnetics (CEM) with a specific focus on solving Maxwell's Equations over a 2D waveguide cross-sections with very high accuracy. This problem has practical engineering applications for the design of waveguides; however, it is used here as a research tool for exploring and evaluating improvements upon FEM.
+ A solution is expressed in terms of a weighted superposition of all the functions in the Domain such that the PDE is satisfied along with some continuity and boundary conditions. Some common applications include the Navier-Stokes equations which characterize the behavior of fluids, Schrödinger's equation which governs the evolution of quantum systems, and Maxwell's Equations which are a macroscopic description of essentially all Electromagnetic phenomena. The ability to accurately and efficiently model these differential equations and others is imperative to the success of many engineering projects and scientific endeavors. 
 
-Although its initial use case was domain specific, `FEM_2D`'s functionality extends easily to other domains using an API based on Rust `Traits` (which are like C++20 `Concepts` or Java `Interfaces`). Traits are a highly expressive form of genericism which allow functions to act on any data structure so long as it implements some shared functionality. In the case of solving PDEs with FEM, that generic functionality is the ability to compute an integral over an expression of two overlapping basis functions. As such, the library's entire functionality can be leveraged against a given PDE problem by implementing FEM_2D's `Integral` trait to express the variational form of the problem of interest. 
+The `FEM_2D` library is a powerful FEM simulation tool implemented in Rust. It was designed within the domain of Computational Electromagnetics (CEM) with a specific focus on solving Maxwell's Equations over a 2D waveguide cross-sections with very high accuracy. This problem has practical engineering applications for the design of waveguides; however, it is used here as a research tool for exploring and evaluating improvements upon FEM. Particularly, this library employs the Refinement-by-Superposition (RBS) approach to $hp$-refinement. RBS is a novel and relatively unexplored generalization of the more typical Refinement-by-Replacement approach which unlocks a much simpler pathway to enforcing continuity conditions. 
+
[…]
