+++ src/fem_domain/basis.rs
@@ -43,19 +43,19 @@ pub trait HierCurlBasisFnSpace: Clone + Sync + Send + std::fmt::Debug {
     fn tang_d2(&self, n: usize, p: usize) -> f64;
 }
 
-/// A utility structure used to generate and cache [BasisFn]'s during Integration
+/// A utility structure used to generate and cache [HierBasisFn]'s during Integration
 ///
 /// Three important areas of functionality are handled here:
-/// * Convert `BasisSpec`s (abstract basis function definitions on an [Elem]) into [BasisFn]'s (numerical instantiations of basis functions for integration)
+/// * Convert `BasisSpec`s (abstract basis function definitions on an [Elem]) into [HierBasisFn]'s (numerical instantiations of basis functions for integration)
 /// * Handle interpolation between parametric spaces to set up inter-layer integration
-/// * Cache computed [BasisFn]'s to avoid re-computation.
+/// * Cache computed [HierBasisFn]'s to avoid re-computation.
 ///
-pub struct BasisFnSampler<B: BasisFn> {
-    /// Maximum u-directed expansion order. [BasisFn]s Generated by this sampler will be defined up to this order in `i`
+pub struct BasisFnSampler<B: HierBasisFn> {
+    /// Maximum u-directed expansion order. [HierBasisFn]s Generated by this sampler will be defined up to this order in `i`
     pub i_max: usize,
-    /// Maximum v-directed expansion order. [BasisFn]s Generated by this sampler will be defined up to this order in `i`              
+    /// Maximum v-directed expansion order. [HierBasisFn]s Generated by this sampler will be defined up to this order in `i`              
     pub j_max: usize,
-    /// Whether the 2nd derivatives of the [BasisFn]s will be computed. If true, the endpoints [-1, +1] are also defined along both axes!                 
+    /// Whether the 2nd derivatives of the [HierBasisFn]s will be computed. If true, the endpoints [-1, +1] are also defined along both axes!                 
     pub compute_d2: bool,
     /// Gauss Legendre Quadrature points evaluated along u-direction. Defined from (-1 to +1)         
     u_points: Vec<f64>,
@@ -65,14 +65,14 @@ pub struct BasisFnSampler<B: BasisFn> {
     computed: Arc<Mutex<HashMap<BSDescription, Arc<B>>>>,
 }
 
-impl<B: BasisFn> BasisFnSampler<B> {
+impl<B: HierBasisFn> BasisFnSampler<B> {
     /// Construct a Basis Function Sampler with the following parameters:
     ///
     /// * `i_max` : maximum u-directed expansion order
     /// * `j_max` : maximum v-directed expansion order
     /// * `num_u_points` : number of Gauss-Leg-Quad points to define along the u-axis. If `None`: a default value is used
     /// * `num_v_points` : number of Gauss-Leg-Quad points to define along the v-axis. If `None`: a default value is used
-    /// * `compute_2nd_derivs`: whether or not to compute the second derivatives of the [BasisFn]s.
+    /// * `compute_2nd_derivs`: whether or not to compute the second derivatives of the [HierBasisFn]s.
     pub fn with(
         i_max: usize,
         j_max: usize,
@@ -102,7 +102,7 @@ impl<B: BasisFn> BasisFnSampler<B> {
         )
     }
 
-    /// Generate or retrieve a [BasisFn] defined over an [Elem]. Can be defined over a subset of the `Elem`.
[…]
