### 2022-04-29 d574fc5
small api change to galerkin_sample_gep_hcurl
