### 2022-04-29 90c787e
Paper Additions. And fixed problem with max_ortho_basis feature
