+++ paper.md
@@ -43,17 +43,7 @@ Efficiently computing FEM solutions over geometries with sharp edges or stark ma
 
 For research purposes, it is also important that the implementation is straightforward and easy to understand. This way, other researchers can quickly read the code to validate the methodology itself or they can use it as a starting point for additional investigation and software development. This is yet another benefit of the RBS approach, as it greatly simplifies the enforcement of continuity conditions, which is typically the most challenging aspect of an *h*-refinement implementation over quadrilateral or hexahedral elements. 
 
-Thus, we conclude that `FEM_2D`'s RBS implementation gives it a distinct advantage over other FEM libraries such as Deal.II [@dealII93]; specificity as a research package. The succinctness of the continuity enforcement algorithm removes much of the difficulty of implementing new features, and the generic Trait-Based interface makes it easy to explore extensions into other problem domains. 
-
-# Project Goals
-
-`FEM_2D`: is intended to be the following:
-
-* A proof-of-concept for the RBS method and associated research
-* A high quality implementation of the RBS approach to FEM for the purpose of encouraging additional related research
-    * This could involve additions to the `FEM_2D` library to support new features or application domains
-    * It could also simply serve as a starting point, or example implementation, for new RBS based FEM software
-* A powerful solver for the 2D Maxwell Eigenvalue Problem in itself (or any other problems that are implemented by the open-source community)
+Thus, we conclude that `FEM_2D`'s RBS implementation gives it a distinct advantage over other FEM libraries such as Deal.II [@dealII93]; specificity as a research package. The succinctness of the continuity enforcement algorithm removes much of the difficulty of implementing new features. This is a major barrier to entry for contributing to larger and more complex packages. Additionally, the generic Trait-Based interface makes it easy to explore extensions into other problem domains. 
 
 # Features
 
@@ -105,9 +95,7 @@ fn do_some_h_refinements(mesh_file_path: &str) -> Result<Mesh, Box<dyn Error>> {
     Ok(mesh)
 }
 ```
-The following example shows how some of the $p$-refinement methods may be used. Here, the `Mesh` is provided as an argument rather than being loaded from a file. 
-
-$P$-refinements are constructed from the `PRef` Type using a pair of `i8`'s (8-bit signed integers). As such, any element's u- and v-directed expansion orders can be modified independently in either the positive or negaitve direction.
+The following example shows how some of the $p$-refinement methods may be used. Here, the `Mesh` is provided as an argument rather than being loaded from a file. The $p$-refinements are constructed from the `PRef` Type using a pair of `i8`'s (8-bit signed integers). As such, any element's u- and v-directed expansion orders can be modified independently in either the positive or negative direction.
 
 The behavior of these methods is straightforward with the slight caveat that the `global_p_refinement` and `p_refine_with_filter` methods will guard against any refinement pushing an element outside of its valid expansion order range. Specifically, refinements are clamped element-wise to ensure that the final expansion order is in the range [1, 20]. The $p$-refinement methods that can return an error (those followed by a `?` in the example) do not exhibit this behavior. This is in keeping with the design of the $h$-refinement API in the sense that methods with less explicit control are safer, while the more explicit methods allow for failure. 
 
@@ -142,7 +130,6 @@ fn do_some_p_refinements(mesh: &mut Mesh) -> Result<(), PRefError> {
     Ok(())
 }
 ```
-
 
[…]
