### 2022-04-30 c7bd180
new version with small documentation, naming and API changes that came up during paper rewrite
