+++ Cargo.toml
@@ -9,7 +9,7 @@ license = "MIT"
 name = "fem_2d"
 readme = "README.md"
 repository = "https://github.com/jeremiah-corrado/fem_2d/"
-version = "0.2.0"
+version = "0.2.1"
 
 [features]
 default = ["json_export"]
