### 2022-04-30 jeremiah-corrado/fem_2d#2 @jeremiah-corrado
Hi, @YohannDudouit, thanks again for the feedback! Quite a few changes have been made to the crate to address your concerns. I also updated the paper to reflect these changes.

The following is a point-by-point response to your observations/suggestions. Each point either explains how the crate was updated to reflect your suggested change, or explains why the suggested change was not made. There are also a few questions mixed in asking for clarification.

# Structure/Organization/Documentation

•	The module structure has been separated into two high level modules: `fem_domain` and `fem_problem`.  `fem_domain` contains the Domain, Mesh and Basis Function structures (everything needed to set up an FEM Domain). `fem_problem` contains the integration and linear algebra stuff for defining and solving FEM problems. There is still a prelude module with most of the important structures, functions and traits.
•	All the GLQ functionality has been moved into a single [GLQ module]( https://docs.rs/fem_2d/0.2.1/fem_2d/fem_problem/integration/glq/index.html) under the Integration module (with improved documentation).
•	`ShapeFn` has been renamed to [HierCurlBasisSpace]( https://docs.rs/fem_2d/0.2.1/fem_2d/fem_domain/basis/trait.HierCurlBasisFnSpace.html). This is a special case of the more general concept: Basis-Space. The new documentation--along with the docs for [HierCurlBasisFn](https://docs.rs/fem_2d/0.2.1/fem_2d/fem_domain/basis/struct.HierCurlBasisFn.html)--should better explain how to create a custom implementation.
•	The names of the basis function structs/traits have all been changed. There is now a [HierBasisFn]( https://docs.rs/fem_2d/0.2.1/fem_2d/fem_domain/basis/trait.HierBasisFn.html) Trait that will be the parent trait for any other Hierarchical Basis function Types. There is currently no Trait for non-hierarchical basis functions.
•	The new names in the `basis` module are now more descriptive: `HierBasisFn` refers to the actual instantiation over some Elem, while a BasisFnSpace describes the generically-specified functional space used to compute the instantiation.
•	I think the updated documentation is sufficient to clarify the conceptual difference between `Elem` and `Element`
•	I am having trouble understanding the logic behind making this separation. The `Mesh` describes the set of `Elements` and `Elems` that make up the discretization. These structures themselves are composed of `Nodes` and `Edges`. They are also the structures that are subject to *h*-refinement and *p*-refinement. All of them are operating within the same conceptual space and thus fit nicely within the same module.
•	M2D and V2D should be in the space module as part of the broader mesh module because they describe points and transformations in the 2D physical/parametric spaces where elements are defined. These are specialized data structures that do not have any utility to the linear algebra portion of the crate

# On the Design:

•	Similar to the point above, I don't see the benefit in having another mesh Type. Currently the `Mesh` structure describes the set of `Element`s and `Elem`s that make up the FEM Domain. What would be gained by making a conceptual separation between the geometric features (Cells, Faces, Edges, Nodes), and the *hp*-refinement stuff? As far as I understand, there is never a stage of the FEM simulation process where we would want an instantiated definition of the Mesh that does not have the capacity to be refined.
•	I am defining a Degree-of-Freedom as a set of one or more [BasisSpec](https://docs.rs/fem_2d/0.2.1/fem_2d/fem_domain/domain/dof/basis_spec/index.html)s. A `BasisSpec` gives an abstract definition for a Basis Function over an individual `Elem`.  Specifically, it defines the expansion-orders, direction, and classification (Element-Type, Edge-Type, Node-Type). Element-Type basis functions constitute degrees of freedom on their own, as they are zero-valued along the relevant edges. Pairs of Edge-Type functions constitute degrees of freedom, as they are non-zero on one relevant edge, and must share a scalar value to enforce continuity. Sets of four Node-Type functions make up degrees of freedom for the same reason. As such, a DoF describes a unit of the global solution which can take a scalar value. My design may be clearer if you look under the hood at the `BasisSpecGroup` structure in dof.rs (Perhaps this field of DoF should be made public?).
•	There is an ` impl From<SparseMatrix> for Nalgebra::DMatrix<f64> ` in the linalg module. A user could definitely decompose a [GEP](https://docs.rs/fem_2d/latest/fem_2d/fem_problem/linalg/struct.GEP.html) into its individual matrices and use Nalgebra functions to manipulate them. Is this what you mean?
•	I haven’t implemented all these traits; however, I think I’ve put a nice structure in place for those alternate basis function types to be developed over time.
•	Continuity conditions are checked explicitly on line 42 of `galerkin.rs`. Mathematically speaking, continuity is enforced in the design of the `HierCurlBasisFn` struct, and the matching algorithm for the `BasisSpecs`.
•	The `ParBasisFnSampler` and `BasisFnSampler` structs were combined.
•	I think more complex material definitions are something that other researcher can implement if there is interest. My understanding is that functionally defined material properties are only useful for defining meta-materials which are a pretty specialized mathematical tool. The current setup gives a decent amount of freedom, as both material constants are complex-valued.

# On the implementation

•	There is also a ` HierMaxOrtho` Basis Function Space which can be included with the `max_ortho_basis` feature-flag on the nightly compiler.
•	I’ve included the mathematical definition of the [HierPoly](https://docs.rs/fem_2d/0.2.1/fem_2d/fem_domain/basis/hierarchical_basis_fns/poly/struct.HierPoly.html) basis space in the documentation, as well as a reference to the full paper associated with the `HierMaxOrtho` basis space. As such, I think there is sufficient context for someone to implement other basis spaces.
•	There is a [mass operator](https://docs.rs/fem_2d/0.2.1/fem_2d/fem_problem/integration/integrals/inner/index.html). The Poisson operator would be more useful with H(Div) continuity conditions, which are not implemented yet.
•	I think this is also something that other researcher might want to implement if they need it. The libraries module structure is friendly to these types of additions. Linear solvers could sit next to the eigensolvers in the linalg module. For example, a function called `nalgebra_solve_lin` could be added to the `linalg::nalgebra_solve` module. And the accompanying sampling function could sit next to `galerkin_sample_gep_hcurl` in the ‘galerkin’ module, and reuse the [SparseMatrix] (https://docs.rs/fem_2d/0.2.1/fem_2d/fem_problem/linalg/sparse_matrix/struct.SparseMatrix.html) structure
