### 2022-04-30 jeremiah-corrado/fem_2d#3 @jeremiah-corrado
Hi @jeremylt, thank you again for the feedback!

I've made some updates to the error handling strategy, particularly in the *hp*-refinement API.  The [Mesh documentation](https://docs.rs/fem_2d/latest/fem_2d/fem_domain/domain/mesh/struct.Mesh.html) now has a clear list of which errors could occur for each of these methods.

I also modified the paper to hopefully explain my thinking behind the error handling strategy a little bit better. The examples have been updated to use the `?` operator.

W.R.T your [comment](https://github.com/openjournals/joss-reviews/issues/4172#issuecomment-1108691106) in the review discussion, I have to disagree with your notion that all the methods should return a `Result<(), Error>` in order to prevent future breaking changes. In some cases, the method is designed not to err *by definition*, and thus I think it is actually more confusing to encode that possibility into the function header. For example, [global_h_refinement](https://docs.rs/fem_2d/latest/fem_2d/fem_domain/domain/mesh/struct.Mesh.html#method.global_h_refinement) only applies a refinement to the `Elems` that are eligible for *h*-refinement. Thus, *by defnintion* it will never run into any of the `HRefError` cases (unless there is a bug in the implementaiton of Mesh, in which case a `panic` is appropriate). As such, I think a `Result<(), HRefError>` return-type would contradict the definition of the method, which is confusing for users.

Lastly, some new Error types were created for a few other structures and functions in the library, where I had previously been returning a formatted String as the Error Type.
