### 2022-04-30 jeremiah-corrado/fem_2d#4 @jeremiah-corrado
Thank you for the suggestion! I agree that the names of the fields in `BasisFn` were pretty undescriptive. The naming and [documentation](https://docs.rs/fem_2d/latest/fem_2d/fem_domain/basis/struct.HierCurlBasisFn.html) for that structure have been updated now. In that case, the "raw" simply meant "not inverted". Those fields are now called "jac_inv" and "jac".

In many cases the public fields provide information that I think could be useful for developing code around the Crate. For example, I have used the `h_levels` field in [Elem](https://docs.rs/fem_2d/latest/fem_2d/fem_domain/domain/mesh/elem/struct.Elem.html) to develop some refinement algorithms that set the polynomial expansion orders on an `Elem` to be a function of its *h*-refinement levels.

To make a more general comment on the high degree of publicness: as I described in the updated version of the paper, this package is really intended to be used in a research context, so generally speaking, I erred on the side of over-exposing internal data rather than under-exposing.

I will also note that much of the library's documentation has been updated recently, so hopefully other confusing public fields that you may have noticed also received improved documentation, making them more usable. Please let me know if you see other specific cases where a public field is poorly described or seems like it should definitely be private.
