### 2022-04-30 jeremiah-corrado/fem_2d#3 @jeremylt
Like I said in the other thread, I'm personally uncomfortable designing code that is coupled in that way. I understand that you don't want to change the design and you have reasons why you don't. I think documenting the assumptions that must be respected in future development related to the revelant 'impossible to fail' functions would be enough to close this issue. Future contributors who aren't you may not find these assumptions obvious.
