### 2022-04-30 jeremiah-corrado/fem_2d#4 @jeremylt
The documentation updates look good to me.

I've seen two schools of thought on exposing strut members. On one hand, exposing the strut members makes them part of your API and any changes to your internal design break other people's code that uses your crate. On the other hand, writing all the accessors to all the data you want to expose can be tedious. As long as it's an intentional design decision to expose the internals and make those strut members part of the API, it can work fine.
