### 2022-05-02 cce44dc
added unwrap-assumptions to non-erroring methods in Mesh. Addresses #3
