### 2022-05-02 jeremiah-corrado/fem_2d#3 @jeremiah-corrado
Ok that makes sense. I just made those documentation additions to mesh.rs
