### 2022-05-02 jeremiah-corrado/fem_2d#6 @jeremiah-corrado
@jeremylt I've set up CI testing for pushes/pull-requests into the main branch. Its currently set up to check for build errors, and run all of the doc/unit tests (except it does skip the `slepc_problem` unit-test because it relies on the external slepc solver).
