### 2022-05-02 #4172 @jeremiah-corrado
Hello all,

Just a quick update to summarize some recent progress:

* I made significant changes to the library over the last couple weeks to address @YohannDudouit's and @jeremylt's concerns/suggestions.
* @YohannDudouit, I left a detailed [response](https://github.com/jeremiah-corrado/fem_2d/issues/2#issuecomment-1114055480) to your issue. I believe I addressed most of your concerns. For those that I did not, I either left a question or some reasoning for maintaining the initial design. I welcome any further feedback (or rebuttals to my rebuttals)
* @jeremylt and I are making progress towards resolving his issues. Again, I welcome any further feedback on the points that have not been resolved yet
* The paper has been updated to reflect the design changes in the library
* I also published a new version of the library to crates.io, so [this documentation](https://docs.rs/fem_2d/0.2.1/fem_2d/) is up to date (minus some very small changes in the last few commits).

Thanks for all the feedback and discussion so far!
