### 2022-05-09 28cc6d8
fixed Galerkin capitalization in paper
