### 2022-05-13 jeremiah-corrado/fem_2d#2 @jeremiah-corrado
Hi @YohannDudouit, thank you for the additional feedback!
Below we responded to your suggestions and improvements.

# FEM_2D Goals
First, given that we significantly improved the manuscript in the latest version based on the general feedback we received, we wanted to clarify the library's goals, as we agree that the initial draft of the paper was poorly explained and unclear in this regard. For the library's capabilities, we emphasized the core, novel functionality, namely the ability to conduct Refinement-by-Superposition (RBS) *hp*-refinements with theoretically optimal (i.e., exponential) rates of convergence. We also believe strongly in the need for greater replicability in applied mathematics and scientific computing research, which was one of the primary motivating factors in open sourcing this library. We also hope to assist others in the FEM research community to investigate Refinement-by-Superposition methods via our library or through their own implementations based on or inspired by our library. Our intent is not to replace or even compete with the excellent finite element libraries with decades worth of patches, but instead to supply a novel, yet fully usable approach to arbitrary levels of *hp*-refinements, isotropic and anisotropic, which is not supported by existing open-source projects.

In order to support this type of work, we included a light amount of functionality targeting the Maxwell Eigenvalue Problem that can be extended to other problems and applications. However, with our focus on providing a novel approach to Refinement-by-Superposition, the bulk of the library (in terms of breadth of functionality, number of lines of code, and depth of documentation) is located in the Mesh implementation. Or, more specifically, in the *hp*-refinement API, which has distinct advantages over other FEM software due to the underlying Refinement-by-Superposition implementation. With this in mind, the code in the `mesh` module is of primary importance.

We amended this description in the latest version of the paper to clarify our focus. We also agree that minor adjustments to the implementation framework (particularly surrounding integration), which we have described below, can simplify extensions by other researchers/practitioners to their applications of interest and enhance the generalizability and broad applicability of our library for finite element methods.

# Integration
With that aside, we definitely see your point about the generic interface for integration. In its current state, the various layers of Traits and Structures are somewhat convoluted, and it would be much nicer to have an overarching `Integral` Trait that applies to any kind of Basis-Function.

Below, I've given a brief description of some of the current design choices in response to your questions. I will then describe a modified design that aims to address your concerns.

## Functional Style GLQ:
There is a functional-style Gauss-Quad integration function in the [GLQ Module](https://docs.rs/fem_2d/latest/fem_2d/fem_problem/integration/glq/fn.real_gauss_quad.html), which is similar to what you suggested. This function is used in the implementation of the [Integrals](https://docs.rs/fem_2d/latest/fem_2d/fem_problem/integration/integrals/index.html). The difference is that the generically defined Closure is of the type:
```rust
Fn(usize, usize) -> f64
```
This assumes that the basis function has already been evaluated and simply needs to be sampled at a set of indices in parametric space.

Alternatively, per your suggestion, basis functions (or functions of basis functions) could be computed on the fly in a closure like:
```rust
Fn(f64, f64) -> f64
```
However, this choice does not permit pre-computation/memoization of basis-function evaluations (both the [HierPoly](https://docs.rs/fem_2d/latest/fem_2d/fem_domain/basis/hierarchical_basis_fns/poly/struct.HierPoly.html) and MaxOrtho Basis Spaces are somewhat expensive to compute). The [BasisFnSampler](https://docs.rs/fem_2d/latest/fem_2d/fem_domain/basis/struct.BasisFnSampler.html) is responsible for caching computed Basis-Functions during [Galerkin Sampling](https://docs.rs/fem_2d/latest/src/fem_2d/fem_problem/galerkin.rs.html#66).

## Vectorial-Direction and Expansion-Order Arguments:
Furthermore, we included the `q_dir`, `p_dir`, `p_orders`, and `q_orders` arguments in the `integrate` functions, because a [HierBasisFn](https://docs.rs/fem_2d/latest/fem_2d/fem_domain/basis/trait.HierBasisFn.html) pre-computes ALL the Basis Functions for both directions (`BasisDir::U` and `BasisDir::V`) from zero-ordered up to the specified maximum orders `i_max` and `j_max`. (Or, more accurately, it computes sets of vectors for each direction and order, so that sampling a BasisFn at a specific point amounts to a single multiplication operation.)

As such, the current `integrate` functions essentially do the following:
* Take the pre-computed sets of basis functions P and Q
* Select out the vectorial direction and the orders (i, j) for both functions
* Then apply the integration operation (CurlCurl, Mass, etc.)

## Separation of GLQ Weights:
The weight vectors are separate from the basis function evaluation for two reasons:
* As you said, they are not scaled by *h*-refinements, whereas the GLQ points (*x*) are scaled. As such, the `Integral` can keep track of the weights during Galerkin Sampling, while the `BasisFnSampler` scales and sub-samples the GLQ-points for `Elems` of various sizes.
* Additionally, as mentioned with the two different Closure designs above, we have not chosen to involve the GLQ points in the GLQ integration, but instead have pre-evaluted the basis functions using those points, and only need the GLQ-weights and indices at "integration time".

## New Design

The following sample outlines a redesigned integration framework to permit a more straightforward implementation of integral operators for H1 basis-functions (or basis-functions of any other kind). The relevant H1 Traits and Structs are included to show the general structure/orgainization of their implementation.

Most importantly, the new `Integral` Trait more closely resembles the generic mathematical concept of an inner product between two basis functions.

(I've included the `unimplemented!()` macro and some dummy functions as necessary so that you can paste this into an IDE, compile, make changes, etc.)

```rust
use std::marker::PhantomData;

fn main() {
    // setup GLQ points
    let [xu, wu] = glq_points_and_weights();
    let [xv, wv] = glq_points_and_weights();

    // setup a sampler for an H(curl) implementation of PolyBasis
    let mut hc_bs_sampler = HCurlBasisFnSampler::<PolyBasis>::new(xu, xv);

    // sample two basis functions (Elem-specific info is missing in this example, but isn't relevant to the design)
    let basis_p = hc_bs_sampler.sample(BasisDir::U, [1, 2]);
    let basis_q = hc_bs_sampler.sample(BasisDir::U, [3, 0]);

    // setup a Mass integrator
    let mass = Mass::with_weights(wu, wv);

    // compute an integral with the two basis function samples
    // The HCurl implementation of Mass's integral impl is automatically used due to the argument type
    let solution = mass.integrate(basis_p, basis_q);
}

// ---------------------------------------------------------------------------------------------
// Basis Spaces (implementation details are left out, as they are not relevant to this design summary)
// ---------------------------------------------------------------------------------------------

// have traits for each kind of basis function space
pub trait HCurlBasisSpace {}
pub trait H1BasisSpace {}

// implement some specific basis spaces. Overlap is allowed (i.e. there is a PolyBasis impl for both H1 and HCurl)
pub struct PolyBasis {}
impl HCurlBasisSpace for PolyBasis {}
impl H1BasisSpace for PolyBasis {}

pub struct SomeOtherBasisSpace {}
impl HCurlBasisSpace for SomeOtherBasisSpace {}

// ---------------------------------------------------------------------------------------------
// Basis Function Samplers
// ---------------------------------------------------------------------------------------------

// define a sampler specifically for H(Curl) Basis-Functions
pub struct HCurlBasisFnSampler<B: HCurlBasisSpace> {
    xu: Vec<f64>,
    xv: Vec<f64>,
    bs: PhantomData<B>,
}

impl<B: HCurlBasisSpace> HCurlBasisFnSampler<B> {
    pub fn new(xu: Vec<f64>, xv: Vec<f64>) -> Self {
        Self {
            xu,
            xv,
            bs: PhantomData,
        }
    }

    // This is the most important line of this example!
    // The direction and order are specified by the caller here, rather than in the arguments to integrate()
    pub fn sample(&self, dir: BasisDir, orders: [usize; 2]) -> HCurlBFSample {
        unimplemented!()
    }
}

// define a separate sampler specifically for H1 Basis Functions
pub struct H1BasisFnSampler<B: H1BasisSpace> {
    xu: Vec<f64>,
    xv: Vec<f64>,
    bs: PhantomData<B>,
}

impl<B: H1BasisSpace> H1BasisFnSampler<B> {
    pub fn new(xu: Vec<f64>, xv: Vec<f64>) -> Self {
        Self {
            xu,
            xv,
            bs: PhantomData,
        }
    }

    // The other most important line of this example!
    // The direction does NOT have to be provided as an argument for the H1 basis sample
    pub fn sample(&self, orders: [usize; 2]) -> H1BFSample {
        unimplemented!()
    }
}

// A sample of an H(Curl) basis function
// Contains only the information needed to evaluate the basis function (or its derivative)
//    at a specific order (i, j) and with a specific direction
pub struct HCurlBFSample<'a> {
    dir: BasisDir,
    u: &'a Vec<f64>,
    v: &'a Vec<f64>,
    u_d1: &'a Vec<f64>,
    v_d1: &'a Vec<f64>,
}

impl<'a> HCurlBFSample<'a> {
    pub fn from_vectors(
        dir: BasisDir,
        u: &'a Vec<f64>,
        v: &'a Vec<f64>,
        u_d1: &'a Vec<f64>,
        v_d1: &'a Vec<f64>,
    ) -> Self {
        Self {
            dir,
            u,
            v,
            u_d1,
            v_d1,
        }
    }

    pub fn f(&self, [m, n]: [usize; 2]) -> f64 {
        unimplemented!()
    }

    pub fn curl(&self, [m, n]: [usize; 2]) -> f64 {
        unimplemented!()
    }
}

// A sample of an H1 basis function with some specific order
pub struct H1BFSample {}

impl H1BFSample {
    pub fn f(&self, [m, n]: [usize; 2]) -> f64 {
        unimplemented!()
    }
}

// ---------------------------------------------------------------------------------------------
// Integral Trait and Operators
// ---------------------------------------------------------------------------------------------

// The simplified integral trait which can handle pairs of any kind of Basis Funciton
pub trait Integral<BFS> {
    fn integrate(&self, p_basis_fn: BFS, q_basis_fn: BFS) -> f64;
}

// an example integration operator: Mass
pub struct Mass {
    wu: Vec<f64>,
    wv: Vec<f64>,
}

impl Mass {
    pub fn with_weights(wu: Vec<f64>, wv: Vec<f64>) -> Self {
        Self { wu, wv }
    }
}

// The H(curl) implementation of the Mass integral
impl<'a> Integral<HCurlBFSample<'a>> for Mass {
    fn integrate(&self, p_basis_fn: HCurlBFSample, q_basis_fn: HCurlBFSample) -> f64 {
        unimplemented!()
    }
}

// The H1 implementation of the Mass integral
impl Integral<H1BFSample> for Mass {
    fn integrate(&self, p_basis_fn: H1BFSample, q_basis_fn: H1BFSample) -> f64 {
        unimplemented!()
    }
}

// ---------------------------------------------------------------------------------------------
// Utility Stuff
// ---------------------------------------------------------------------------------------------

// dummy GLQ points function
pub fn glq_points_and_weights() -> [Vec<f64>; 2] {
    unimplemented!()
}

// keep the current BasisDir Enum for vectorial direction selection
pub enum BasisDir {
    U,
    V,
}

```
Please let me know if the above redesign would address your concerns. If so, I will work on modifying the relevant portions of the library accordingly.

# Linear Form

The library's module structure is friendly to this addition, and we believe it would be reasonable for other researchers and practitioners to implement this using the current Galerkin Sampling function as an example. The integration redesign above would also make the integration portion of this addition more straightforward, as the same `Integral` Trait could be used (where `f` would be a different Kind of `BasisFnSample`).

# Non-Constant Material Coefficients

Similarly, since the Refinement-by-Superposition approach is agnostic to the forms of the integrands (including non-constant coefficients), we believe this could eventually be added via open source contribution. The methods currently being used to map from an Element's real space to a descendant Elem's parametric space (during integration) could be readily used to map a functional material parameter on an Element to any child Elem during integration.
