### 2022-07-18 #4172 @jedbrown
Thanks @YohannDudouit for your comprehensive review.

@jeremiah-corrado how are you doing? From the perspective of JOSS, we are looking for accuracy. So if you don't want to make major changes to the library, you can adjust the abstract to make clear that it's a library for linear $H(\text{curl})$ problems or whatever is an accurate statement of what is currently supported. Of course software can hypothetically be extended to do lots of things, but if there isn't an interface with more than one implementation showing how it works, it's unlikely to attract developers to actually do that work (and exaggerated claims are likely to turn off people who might otherwise be capable of doing such work). So please just let us know what you plan to do and don't hesitate to ask questions.

As an aside, I notice that the [quadrature interface](https://docs.rs/fem_2d/latest/fem_2d/fem_problem/integration/glq/fn.real_gauss_quad.html) handles only weights, leaving the points to the user to index manually. This `Fn(usize, usize) -> f64` is an unusual interface in mathematical software; usually the quadrature routine would be aware of the points and thus the integrand would be `Fn(f64, f64) -> f64`, which is easier to use and performs just as well.
