### 2022-08-16 #4172 @jeremiah-corrado
Hi @jedbrown, sorry for the delay in responding to your earlier message!

I plan to rewrite parts of the paper and introductory documentation to reflect the `H(curl)` selectiveness of the library. That should happen this coming weekend.

> As an aside, I notice that the [quadrature interface](https://docs.rs/fem_2d/latest/fem_2d/fem_problem/integration/glq/fn.real_gauss_quad.html) handles only weights, leaving the points to the user to index manually. This Fn(usize, usize) -> f64 is an unusual interface in mathematical software; usually the quadrature routine would be aware of the points and thus the integrand would be Fn(f64, f64) -> f64, which is easier to use and performs just as well.

This was done for performance reasons, as it allows the precomputed basis-function values to be repeatedly called from memory rather than being recomputed. Yohann had a similar concern, which I responded to in more detail here: https://github.com/jeremiah-corrado/fem_2d/issues/2#issuecomment-1126263955
