### 2022-08-21 0c11e8f
Paper updates - Focused functionality claims around H(curl) conformgin problems per JOSS reviewer feedback
