### 2022-08-21 52ae739
Paper updates - Focused functionality claims around H(curl) conformgin problems per JOSS reviewer feedback
