### 2022-08-21 9bf1cae
cleaned up readme.md. focused functionality claims around H(curl) problems.
