### 2022-09-18 #4172 @jeremiah-corrado
Hi @jedbrown, I made some updates to the paper a few weeks ago that reduced the claimed scope of the library to the H(curl) functionality that is currently supported. I also make it clear that extending the library into other domains would be more intensive than creating custom implementations of the basis-function and integral traits. I believe these changes should address the remaining issues in the review.

Please let me know what the next steps are. Thank you!
