### 2022-10-25 #4172 @YohannDudouit
Hi @jedbrown, @danielskatz, and @jeremiah-corrado ,
I apologize for the slow answer. My remaining main issue is with the "Statement of need", I think the use of "efficient"/"efficiently" is confusing and misleading, and I am also not comfortable with the claim of "distinct advantage over Deal.II". If I understand correctly the statement of need, the use of "efficient" means "ease of implementation" and also relates to the ability to perform hp-adaptivity.
The ability to perform hp-adaptivity is not exclusive to `FEM_2D`, and in that regard the RBS approach uses more resources (in particular memory) than a "constrained-node refinement" approach (used by `Deal.II` for instance). Unless some sort of performance comparison between `FEM_2D` and `Deal.II` is provided I would suggest to avoid the claims over `Deal.II`, and just mention `Deal.II` as one of the FEM library using a "constrained-node refinement" approach in the introduction (maybe line 33?).
Regarding the "ease of implementation", I have a mixed feeling about these claims, if I understand how this can be an important point for the "efficiency" of developers, I think this is misleading to categorize this under "Efficiently computing FEM solutions" (line 48). Moreover, if the RBS approach allows easier implementation of continuity conditions when compared to a "constrained-node refinement" approach, as mentioned in previous conversations, the public API is relatively polluted by implementation details; for instance the [quadrature interface](https://github.com/openjournals/joss-reviews/issues/4172#issuecomment-1216679829) uses a public API close to the implementation for "performance reasons", when these implementation details could be hidden from the user by providing a public API closer to the mathematics and keeping these optimization details in a private API. Therefore, I would suggest to clarify the different pros and cons of the RBS approach over a "constrained-node refinement" approach, and explain how this can benefit FEM solutions that require hp-adaptivity. In particular, the end of the introduction (line 33-41) explains "what" is different in the RBS from a "constrained-node refinement" approach, but leaves the reader wondering "so what". In particular, I would suggest clarifying why one should care about these RBS features promoted in the introduction, i.e. (1) why should one care to have a "forest" of elements instead of a "flat" collection of elements, and (2) what is the use of integrating functions defined on different layers of element trees (mixed mesh resolution FEM?).
@jeremiah-corrado Please feel free to discuss these points here before modifying the paper.
