### 2022-10-30 #4172 @jeremiah-corrado
Hi @YohannDudouit, thanks for the additional feedback. It sounds like we are mostly on the same page w.r.t what kinds of claims the paper should be making; however, some of my wording and explanation probably need to be improved.

TL;DR version of the following: I think the way you were interpreting the paper conflated some background info about per-dof-efficiency with the claims I was making about the benefits of the RBS method (over other efficient methods). There are several places — particularly in the statement of need — where the wording can be improved to alleviate this confusion. Let me know what you think about the following clarifications:

---

>  I think the use of "efficient"/"efficiently" is confusing and misleading

In the "Statement of Need" section, I am using the word "efficiently" to mean something very specific: iteratively converging on a solution at an exponential rate w.r.t. the number of degrees of freedom.

The discussion of efficiency is strictly there to motivate the use of *hp*-refinement in general (as opposed to using *p*-refinement alone). I am not intending to make any claims that the library is somehow more efficient than other methodologies or libraries. I am saying that it provides convergence rates that are in the same class as other FEM libraries like Deal.II. Namely, exponential convergence as opposed to algebraic.

The goal was to convey this idea without digging into the math of theoretical convergence rates, but perhaps I should include a more rigorous definition of "efficient/efficiently" in this section? Or use a different word all together?

>  I am also not comfortable with the claim of "distinct advantage over Deal.II".

The distinct advantage over Deal.II is that the `Mesh` data structure in the library imposes fewer constraints on *h*-refinement. Namely, anisotropic *h*-refinements are allowed for Continuous Galerkin problems, and there is no limit on the number of hanging nodes along each edge.

Deal.II, by contrast, only allows anisotropic refinement of hexahedral elements for Discontinuous-Galerkin problems.And even then, it limits refinements to one hanging-node per edge. See this example from their documentation: https://www.dealii.org/developer/doxygen/deal.II/step_30.html.

I can change the phrasing from "distinct advantage" to "an advantage"? Or perhaps state that this is an advantage over the constrained-nodes approach in general instead of discussing Deal.II?

> The ability to perform hp-adaptivity is not exclusive to FEM_2D, and in that regard the RBS approach uses more resources (in particular memory) than a "constrained-node refinement" approach (used by Deal.II for instance). Unless some sort of performance comparison between FEM_2D and Deal.II is provided I would suggest to avoid the claims over Deal.II, and just mention Deal.II as one of the FEM library using a "constrained-node refinement" approach in the introduction (maybe line 33?)

The paper doesn't make any performance claims over Deal.II. It certainly does not claim that `fem_2d` is the only library that implements *hp*-adaptivity. I'm not sure where you are seeing that?

I'd be happy to include a few sentences about the tradeoffs between constrained-nodes and RBS as it pertains to: memory consumption, implementation complexity (lines of code, not big-O), etc.

> Regarding the "ease of implementation", I have a mixed feeling about these claims, if I understand how this can be an important point for the "efficiency" of developers, I think this is misleading to categorize this under "Efficiently computing FEM solutions" (line 48)

Again, line 48 is only justifying the usefulness of *hp*-refinement as such. This is intended to establish background information, particularly for the reader who is wondering why someone might be concerned with *hp*-refinement at all.

In other words, the logical progression of this section was intended to be something like:
- Efficiently solving FEM problems is important
- For hard problems, this requires *h* and *p* refinement
- `fem_2d` is _a_ library that implements *hp*-refinement
	- In fact, its primary intelectual contribution is its *hp*-refinement API
- `fem_2d` might be of particular interest (over some other libraries that have *hp*-adaptivity) because RBS also gives it other interesting features like:
	- anisotropic *h*-refinements
	- n-irregularity
	- a succinct and easy to follow implementation of the Mesh data structure and associated *hp*refinement code

> mentioned in previous conversations, the public API is relatively polluted by implementation details; for instance the quadrature interface uses a public API close to the implementation for "performance reasons", when these implementation details could be hidden from the user by providing a public API closer to the mathematics and keeping these optimization details in a private API.

I agree that the portion of the API that pertains to integration has a lot of implementation details. However, in a response to one of your previous messages (the first few paragraphs of: https://github.com/jeremiah-corrado/fem_2d/issues/2#issuecomment-1126263955), I was really trying to convey the idea that `fem_2d`'s primary contribution is it's `Mesh` data structure and *hp*-refinement API. The portions of the code related to integration and Maxwell's Equation are really there to prove the validity of the method. The structure of the associated research is very similar in the sense that RBS as a methodology is the primary focus, and Maxwell's Eigenvalue Problem is used as an avenue to experiment with it.

I also think the structure of this paper conveys the same idea. The "hp-Refinement API" section is the longest, and is the focus of the introduction, conclusion and statement of need. The "Problem Formulation and Solution" section, by contrast, is much shorter and states explicitly that it is meant as an example of how the *hp*-refinement API might be employed. See line 124.

Please let me know if you disagree that this is the central focus of the paper/documentation, and I can try to make it more clear.

> In particular, the end of the introduction (line 33-41) explains "what" is different in the RBS from a "constrained-node refinement" approach, but leaves the reader wondering "so what". In particular, I would suggest clarifying why one should care about these RBS features promoted in the introduction, i.e. (1) why should one care to have a "forest" of elements instead of a "flat" collection of elements, and (2) what is the use of integrating functions defined on different layers of element trees (mixed mesh resolution FEM?).

I agree, this leaves the reader hanging a bit. I can add some more detail/justification to that paragraph.
