### 2022-11-15 #4172 @jedbrown
Thanks for the discussion here. I just want to note that "novelty" per se is not a requirement for JOSS. You have a package written in Rust that implements an hp-adaptive FE method. It doesn't need to have unique algorithmic capability. It does need to compare to related software, and that comparison needs to accurately depict those other packages. If there are kinds of tasks that are simpler or safer with your package, you might mention those. For comparison to MFEM, @YohannDudouit could perhaps mention some common types of user errors and you could assess whether those are possible in FEM_2D. But the bottom line is that JOSS recognizes value in diversity so don't try to force a "better than" narrative.

BTW, you may want to mention [Fenris](https://docs.rs/fenris/), which is a finite element (not hp-adaptive) package written in Rust.
