### 2022-12-14 #4172 @jedbrown
@jeremiah-corrado Will you be able to make revisions based on the comments above, which I think are mostly about the comparison with related work. (Of course you can choose to make code changes too, but that shouldn't be necessary for acceptance.)

@YohannDudouit Could you please revisit your checklist and consolidate what you think remains as necessary changes for acceptance, acknowledging that JOSS's criteria does not insist that the capability be "better" than existing hp-adaptive packages like Deal.II and MFEM.
