### 2022-12-14 #4172 @YohannDudouit
@jedbrown I am mostly okay with the current state. I do not agree that algorithms are the main reason for code complexity, I think the reality is far more complex; my experience is that the number of lines of code depends much more on design choices, programming experience of the developers, code architecture, than the actual algorithms being implemented. For instance the hp-capability in MFEM is currently limited by the serial/parallel design, and has little to do with the hp-algorithms as far as I know. Anyway, I am not requesting to prove the superiority of FEM_2D over Deal.II and MFEM, but I find the claim over Deal.II somehow inappropriate, difficult to grasp for the reader, and not very relevant; if you think this is fine then we can move on.
