### 2022-12-20 #4172 @jeremiah-corrado
@jedbrown Yeah, sorry for the delay, I'll be making changes this weekend. I think I'll remove any mention of other FEM libraries and simply explain what fem_2d has to contribute.
