### 2022-12-20 #4172 @jedbrown
Okay, thanks. Note that you still have to discuss related work (so please keep reference to hp-adaptive FE libraries), you just don't have to claim greater capability. Just try to give the reader an accurate picture.
