### 2022-12-26 617808b
paper alterations per feedback
