+++ paper.bib
@@ -1,4 +1,13 @@
 
+@article{corrado:2022,
+    title={An Adaptive Anisotropic hp-Refinement Algorithm for the 2D Maxwell Eigenvalue Problem},
+    author={Corrado, Jeremiah and Harmon, Jake and Notaros, Branislav},
+    year={2022},
+    month={Apr},
+    url={https://www.techrxiv.org/articles/preprint/An_Adaptive_Anisotropic_hp-Refinement_Algorithm_for_the_2D_Maxwell_Eigenvalue_Problem/19636770},
+    doi={10.36227/techrxiv.19636770.v1}
+}
+
 @article{corrado:2021, 
     title={A Refinement-by-Superposition Approach to Fully Anisotropic hp-Refinement for Improved Efficiency in CEM}, 
     url={https://www.techrxiv.org/articles/preprint/A_Refinement-by-Superposition_Approach_to_Fully_Anisotropic_hp-Refinement_for_Improved_Efficiency_in_CEM/16695163/1}, DOI={10.36227/techrxiv.16695163.v1}, 
