### 2022-12-26 #4172 @jeremiah-corrado
Hello @YohannDudouit, I made some updates to the paper that I believe will address your concerns: https://github.com/jeremiah-corrado/fem_2d/commit/617808b932eceb7b382dffb157a7b63b4834f1cb. Please let me know what you think. Thanks!
