### 2023-01-11 #4172 @YohannDudouit
@jeremiah-corrado I think the manuscript is very close (besides the remark I did on your commit). Though, I don't think you discuss related work as requested by @jedbrown.
