+++ paper.md
@@ -39,17 +39,19 @@ In addition to the centrally important *hp*-refinement functionality, `FEM_2D` i
 
 # Statement of Need
 
-Efficiently computing FEM solutions over geometries with sharp edges or stark material discontinuities necessitates *hp*-refinement (whether isotropic or anisotropic) because these situations tend to introduce multi-scale solution behavior which is challenging to model with pure *p*- or pure *h*-refinements. This motivates the use of solvers with combined *hp*-refinements where an initially coarse mesh is progressively refined both spatially and in terms of polynomial expansion order. Within this framework, at each refinement step, small-scale behavior is localized into smaller elements using *h*-refinement, and modelling  of higher-order behavior is improved using *p*-refinements. When intelligently applied to problems with sharp or singular solutions, combined *hp*-refinements can achieve exponential rates of convergence, where *h*- or *p*-refinements alone would only be able to achieve algebraic convergence [@harmon:2021]. Like other FEM libraries such as Deal.II [@dealII93], `FEM_2D` aims to expose a rich *hp*-refinement API that can be used to iteratively solve challenging FEM problems in such a manner.
+Efficiently computing FEM solutions over geometries with sharp edges or stark material discontinuities necessitates *hp*-refinement (whether isotropic or anisotropic) because these situations tend to introduce multi-scale solution behavior which is challenging to model with pure *p*- or pure *h*-refinements. This motivates the use of solvers with combined *hp*-refinements where an initially coarse mesh is progressively refined both spatially and in terms of polynomial expansion order. Within this framework, at each refinement step, small-scale behavior is localized into smaller elements using *h*-refinement, and modelling  of higher-order behavior is improved using *p*-refinements. When intelligently applied to problems with sharp or singular solutions, combined *hp*-refinements can achieve exponential rates of convergence, where *h*- or *p*-refinements alone would only be able to achieve algebraic convergence. `FEM_2D` exposes an *hp*-refinement API that can be used to achieve such results [@harmon:2021].
 
 Within the class of *hp*-refinements, the addition of anisotropic *hp*-refinements (over isotropic ones) presents a larger capacity for solution efficiency, as small-scale behavior can be targeted more directly and ineffectual Degrees of Freedom (DoFs) can be left out of the system [@corrado:2021]. In other words, these directionally specific refinements can reduce the introduction of superfluous entropy into the system by directly targeting inaccuracies that are specific to only one direction. The resultant improvement in per-DoF efficiency can be used to reduce the memory requirements for a given solution accuracy or to increase the accuracy achievable with a given amount of memory. Thus, providing a feature-rich anisotropic *hp*-refinement API is an additional goal of the `FEM_2D` library.
 
-For research purposes, it can also be helpful to design software libraries that are straightforward to use and understand. As such, we note that `FEM_2D` is available on Rust's package manager Cargo, making it straightforward to download, compile and run using only a few commands. It can also be included as a dependency in any Rust project to develop new code on top of the library. Additionally, the Refinement by Superposition (RBS) approach that underpins `FEM_2D` lends itself to a desirable level of simplicity with respect to its *h*-refinement implementation. This is because RBS is designed such that continuity conditions between neighboring elements are enforced by construction — no explicit handling of hanging nodes is required. As such some of the typical difficulties with implementing *h*-refinements over quadrilateral elements for H(curl) or H(div) conforming boundary conditions are avoided entirely. Our hope is that the straightforward nature of the RBS approach will allow other researchers to easily contribute to `FEM_2D`, or to use it as a starting point for software development in adjacent research.
[…]
