### 2023-01-26 #4172 @jeremiah-corrado
Hi @jedbrown and @YohannDudouit, I've made some updates with a brief related-work paragraph in the Statement of Need section: https://github.com/jeremiah-corrado/fem_2d/commit/dcc6c2116a07b0b16848d9c7fc2bf4c895d12c19. Please let me know if this addresses your concerns. Thanks!
