### 2023-02-03 #4172 @YohannDudouit
@jeremiah-corrado Thank you for your patience and devotion to improve your code and publication. I am very satisfied with the current state of your publication, I think it highlights clearly the strengths of your RBS approach. I hope to see your library grow and extend its functionalities to 3D and more PDE problems.

@jedbrown As stated above I am fully satisfied with the article, and all my concerns have been addressed.
