### 2023-02-04 #4172 @jedbrown
Thanks, @YohannDudouit. Could you tick the remaining boxes in [your checklist](https://github.com/openjournals/joss-reviews/issues/4172#issuecomment-1065597022)?

@jeremiah-corrado Thanks for your work. Just some copy edits and then I'll ask you to archive.
- [ ] Review these https://github.com/jeremiah-corrado/fem_2d/pull/10
- [ ] The first paragraph opens in a super general setting, but is suddenly pure `system of linear equations`. FE is used for a lot of nonlinear problems (sometimes without linearization at any stage) so the problem space should either be qualified or linear algebra could be identified as a tool. Note that the next paragraph mentions nonlinear problems, which as I understand it, FEM_2D doesn't explicitly handle. Maybe these should be cut.
- [ ] The hp "requirement" is relative to accuracy tolerances. Many engineering problems in high-contrast materials are solved sufficiently accurately using relative low fixed order elements (indeed, this is the rule, and hp adaptivity is rare in production), but there are examples (especially in frequency domain problems) that need such adaptivity to reach engineering tolerances. Asymptotically fast convergence certainly requires it, but applied science and engineering doesn't care about asymptotic convergence.
- [ ] It looks like functions like `execute_p_refinements` don't need ownership and thus could take a slice instead of a `Vec`. (No need to do anything about this, but it occurred to me while reading.)
- [ ] I see that `U` and `V` feature prominently in the design. Can fem_2d run on an unstructured mesh (to discretize an irregular domain or conform to curved surfaces)? What does `U` and `V` mean in that case? Does the caller need to figure out the principle orientation of the element to make an informed choice about anisotropic refinement? (Nothing to do, but maybe you'll want to clarify.)
- [ ] Perhaps you'd like to newtype the expansion order so that ranges are validated at that time rather than passing `u8`/`i8` and needing to validate within a mutation interface? OTOH, this may not be possible if you have elements where 1 is not valid, for example. Note that 0 is a valid order for DG and mixed formulations (with "pressure" in Stokes and H(div) problems) so the rules may need to be more context-specific than they are now.
- [ ] The $B_0$ in equation 1 should be just $B$ according to the definition that follows?
- [ ] `HierPoly` is referred to as using "exponential functions", but it looks like just polynomials in the code?
- [ ] `H(Curl)` and `H(curl)` have inconsistent capitalization. Also `H(Div)` and `H(div)`. Lowercase is more typical, but they at least need to be consistent.
- [ ] We should let https://github.com/openjournals/inara/pull/26 merge to fix the ligature problem with `>>` in the code.

- [ ] I see that running the tests modifies files that are committed to the repository. Visualizing the two files, the domains and solutions are very different.
$ git diff --stat=90
 test_output/mesh_b_fields.vtk | 2109 ++++++++++++++++++++++++++++++++++++++++++++++++---
 1 file changed, 2015 insertions(+), 94 deletions(-)
```
