### 2023-02-16 96a8262
JOSS paper copy edits (#10)
