### 2023-02-16 #4172 @jeremiah-corrado
Hi @jedbrown, thanks for the feedback and copy edits! Your PR was merged, and I believe I've resolved most of your concerns from above in: https://github.com/jeremiah-corrado/fem_2d/commit/ee15513aae873c566e2686927030960f2538fbea

> I see that running the tests modifies files that are committed to the repository. Visualizing the two files, the domains and solutions are very different.

Those must have been from an old test run or something. I decided to `gitignore` the contents of that directory.

> Perhaps you'd like to newtype the expansion order so that ranges are validated at that time rather than passing u8/i8 and needing to validate within a mutation interface?

I'm not sure what you mean by this.
