### 2023-02-16 #4172 @jedbrown
Cool! On your last point, you spend a fair amount of time discussing this point so evidently think it's important:

> Here, both methods can return an error, as it is possible to specify an invalid expansion order. These methods take a length-two array of `u8`'s (8-bit unsigned integers), and thus preemptively remove the possibility of setting negative expansion orders, however, they still Err on expansion orders that are zero or too large.

An alternative is to have `struct Order(u8)` and `let o = Order::try_new(0)?` return `Err` so that mutation interfaces can avoid error in such cases. Generally, mutation failure is harder to recover from because it's messy to rely on the intermediate state (and some mutation interfaces have move semantics).
