### 2023-03-03 #4172 @jeremiah-corrado
Hi @jedbrown, I agree with you that the API is much nicer if erroneous expansion orders are caught earlier.

I created a PR to implement this idea: https://github.com/jeremiah-corrado/fem_2d/pull/11/files.

Let me know if this is what you were thinking and I'll merge it.
