+++ src/fem_domain/domain/mesh.rs
@@ -18,7 +18,7 @@ use elem::{Elem, ElemUninit};
 use element::{Element, Materials};
 use h_refinement::{HRef, HRefError};
 use node::Node;
-use p_refinement::{PRef, PRefError};
+use p_refinement::{PRef, PRefError, Orders};
 use space::{ParaDir, Point};
 
 use super::IdTracker;
@@ -612,10 +612,10 @@ impl Mesh {
     /// let mut mesh = Mesh::unit();
     /// mesh.h_refine_elems(vec![0], HRef::T).unwrap();
     /// mesh.set_expansion_orders(vec![
-    ///     (1, [1, 1]),
-    ///     (2, [3, 3]),
-    ///     (3, [2, 7]),
-    ///     (4, [4, 1]),
+    ///     (1, Orders::new(1, 1)),
+    ///     (2, Orders::new(3, 3)),
+    ///     (3, Orders::new(2, 7)),
+    ///     (4, Orders::new(4, 1)),
     /// ]).unwrap();
     ///
     /// let [max_u_order, max_v_order] = mesh.max_expansion_orders();
@@ -1248,8 +1248,8 @@ impl Mesh {
     ///
     /// let mut mesh = Mesh::unit();
     /// mesh.global_h_refinement(HRef::T);
-    /// mesh.set_expansion_on_elems(vec![1], [MAX_POLYNOMIAL_ORDER - 2, 2]);
-    /// mesh.set_expansion_on_elems(vec![2], [3, 5]);
+    /// mesh.set_expansion_on_elems(vec![1], Orders::try_new(MAX_POLYNOMIAL_ORDER - 2, 2).unwrap());
+    /// mesh.set_expansion_on_elems(vec![2], Orders::try_new(3, 5).unwrap());
     ///
     /// // apply a global p-refinement
     /// mesh.global_p_refinement(PRef::from(3, -3));
@@ -1525,21 +1525,20 @@ impl Mesh {
     /// mesh.global_h_refinement(HRef::T);
     /// mesh.global_h_refinement(HRef::T);
     ///
-    /// mesh.set_global_expansion_orders([4, 5]).unwrap();
+    /// mesh.set_global_expansion_orders(Orders::new(4, 5));
     /// for elem in mesh.elems.iter() {
     ///     assert_eq!(elem.poly_orders.ni, 4);
     ///     assert_eq!(elem.poly_orders.nj, 5);
     /// }
     /// ```
-    pub fn set_global_expansion_orders(&mut self, orders: [u8; 2]) -> Result<(), PRefError> {
-        self.set_expansion_orders(self.elems.iter().map(|elem| (elem.id, orders)).collect())
+    pub fn set_global_expansion_orders(&mut self, orders: Orders) {
+        self.set_expansion_orders(self.elems.iter().map(|elem| (elem.id, orders)).collect()).unwrap()
     }
 
     /// Set the expansion orders on a list of [Elem]s by their IDs
     ///
     /// # Returns an `Err` if
     /// * Any of the `elem_id`'s does not exist
-    /// * The `orders` are outside the valid range: `[1, MAX_POLYNOMIAL_ORDER]`
     /// * There are duplicate `elem_id`s
     ///
     /// # Example
@@ -1552,7 +1551,7 @@ impl Mesh {
     /// assert_eq!(mesh.elems[0].poly_orders.nj, 1);
     ///
     /// // set expansion orders to `[6, 3]`
-    /// mesh.set_expansion_on_elems(vec![0], [6, 3]).unwrap();
+    /// mesh.set_expansion_on_elems(vec![0], Orders::new(6, 3)).unwrap();
     ///
[…]
