+++ Cargo.toml
@@ -9,7 +9,7 @@ license = "MIT"
 name = "fem_2d"
 readme = "README.md"
 repository = "https://github.com/jeremiah-corrado/fem_2d/"
-version = "0.2.1"
+version = "0.2.2"
 
 [features]
 default = ["json_export"]
