### 2023-04-20 #4172 @jeremiah-corrado
Great, I believe those tasks are all complete:
* Release Version 0.2.2: https://github.com/jeremiah-corrado/fem_2d/releases/tag/v0.2.2
* Zenodo DOI: https://doi.org/10.5281/zenodo.7849878
