+++ paper.bib
@@ -1,6 +1,6 @@
 
 @article{corrado:2022,
-    title={An Adaptive Anisotropic hp-Refinement Algorithm for the 2D Maxwell Eigenvalue Problem},
+    title={An Adaptive Anisotropic hp-Refinement Algorithm for the 2D {M}axwell Eigenvalue Problem},
     author={Corrado, Jeremiah and Harmon, Jake and Notaros, Branislav},
     year={2022},
     month={Apr},
