### 2023-04-21 #4172 @danielskatz
@jeremiah-corrado - I've suggested some changes for language and to fix the warnings in https://github.com/jeremiah-corrado/fem_2d/pull/12. You also might want to define/explain h and p in the fourth paragraph of the paper (or before that).
