### 2023-04-21 #4172 @danielskatz
@jedbrown - As a note, this paper now seems far too long for a JOSS paper, and I think that a lot of it probably should be in the documentation, but its not worth changing at this point. I don't think it was so long when it was first submitted, so I assume additions were made during the review process. If I'm wrong and it started out this long, sorry. But if I'm right, please consider this in future reviews, where some of the details in the paper be in the documentation instead.
