### 2023-04-21 ca29cbc
adding h- and p-refinement description to paragraph four of paper
Signed-off-by: Jeremiah Corrado <jeremiah.corrado@hpe.com>
