### 2023-04-21 bfa1a5b
minor changes in paper and bib (#12)
* minor changes in paper
* minor change in bib
