### 2023-04-21 #4172 @jeremiah-corrado
Thanks @danielskatz, I merged your PR and added a brief description of *h*- and *p*-refinement to paragraph four.
