### 2023-04-21 #4172 @danielskatz
Congratulations to @jeremiah-corrado (Jeremiah Corrado) and co-authors on your publication!!

We couldn't do this without your voluntary effort

(I see that the DOI isn't yet resolving, so I'll keep this issue open until it does, which hopefully should be in the next few hours at the latest)
